### 2026-03-21 #10237 @danielskatz
👋 @Axect  - thanks for your submission.

I saw a duplicate submission of this, so I removed the first one; this is the second.

While I look for an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. An editorialbot command needs to be the first entry in a new comment.
