### 2026-03-22 8489197
PAPER: Add missing DOI/ISBN to paper.bib
