+++ paper/paper.bib
@@ -162,7 +162,8 @@
   title     = {Numerical Recipes: The Art of Scientific Computing},
   edition   = {3rd},
   publisher = {Cambridge University Press},
-  year      = {2007}
+  year      = {2007},
+  isbn      = {978-0-521-88068-8}
 }
 
 % ===== Python ecosystem references =====
@@ -223,13 +224,12 @@
   doi     = {10.3389/fceng.2021.758090}
 }
 
-@misc{adtrait,
-  author        = {Liang, Chen and Wang, Qian and Xu, Andy and Rakita, Daniel},
-  title         = {ad-trait: A Fast and Flexible Automatic Differentiation Library in {Rust}},
-  year          = {2025},
-  eprint        = {2504.15976},
-  archiveprefix = {arXiv},
-  url           = {https://arxiv.org/abs/2504.15976}
+@inproceedings{adtrait,
+  author    = {Liang, Chen and Wang, Qian and Xu, Andy and Rakita, Daniel},
+  title     = {ad-trait: A Fast and Flexible Automatic Differentiation Library in {Rust}},
+  booktitle = {2025 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  year      = {2025},
+  doi       = {10.1109/iros60139.2025.11245908}
 }
 
 @software{autodiff_elrnv,
