### 2026-03-22 #10237 @Axect
@danielskatz Thanks for the guidance! I've updated the references:

- **`ad-trait`**: Updated from an arXiv preprint to the published IROS 2025 conference paper with DOI `10.1109/iros60139.2025.11245908`.
- **Numerical Recipes**: The suggested DOI (`10.1080/00401706.1987.10488304`) is actually a 1987 book review of the 1st edition, not the 3rd edition (2007) we cite. I've added the ISBN (`978-0-521-88068-8`) instead, as this book does not appear to have a DOI.
- **Apache Parquet**: The suggested DOI (`10.1007/978-1-4842-2199-0_8`) refers to a book chapter about Parquet, not the software itself. As an open-source software project, Apache Parquet does not appear to have a DOI.
