### 2026-04-01 #10237 @jbytecode
@Axect - Hi, thank you for submitting your work to JOSS. I'm the handling editor of this submittion.

I've just cloned the project, tested in my Linux machine, and all of the tests have been passed. It's also a good and comprehensive numerical computation library with nice implementations of R, Julia-ish syntax. (zeros(n, m) reminded me the Julia function). Nice work.

We need to assign at least two reviewers. Do you have any suggestions for suitable reviewers? You can also use the link `list of people` which is located on top of this issue page. You can search and filter people with their experience, subject and favorite language. Please mention their github handles without using the `@` character to avoid unwanted notifications and emails.

You can also suggest people out of the list. You can provide me their email and github handles.

Thank you in advance.
