### 2026-04-02 #10237 @Axect
@jbytecode - Hi, thank you for testing the project and for the kind words! You're right, many of the API conventions were indeed inspired by Julia and R (e.g., zeros, seq, cbind), as I wanted to make the transition to Rust as natural as possible for scientists coming from those languages.

Here are my suggestions for reviewers:

1. **jeremylt** (Jeremy L Thompson, jeremy@jeremylt.org)
2. **ariostas** (Andres Rios Tascon, ariostas@princeton.edu)
3. **jonaspleyer** (Jonas Pleyer, jonas.pleyer@fdm.uni-freiburg.de)

All three have experience reviewing Rust-related JOSS papers and have no conflicts of interest with this submission.

Thank you!
