### 2026-04-02 #10237 @ariostas
@jbytecode I'd be happy to review it
