### 2026-04-02 #10237 @jbytecode
@jeremylt  - Thank you for the response! Hope we can find a chance to work together in future works.

@jonaspleyer - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

https://joss.readthedocs.io/en/latest/review_criteria.html
