### 2026-04-02 #10366 @editorialbot
[REVIEW]: Peroxide: A Batteries-Included Numerical Computing Library for Rust
**Submitting author:** @axect (<a href="http://orcid.org/0009-0000-4229-2935">Tae-Geun Kim</a>)
**Repository:** https://github.com/Axect/Peroxide
**Branch with paper.md** (empty if default branch): features/joss-paper
**Version:** v0.41.0
