### 2026-04-02 #10366 @jbytecode
@ariostas, @jonaspleyer - Hi, this is the review issue. Please type

`@editorialbot generate my checklist`

to generate your check list. This command will create a task list for each single reviewer.

@Axect - Please keep in monitoring the issue. Our reviewers can eventually report their comments on both the repo and the paper.

Thank you all in advance.
