### 2026-04-04 #10366 @jonaspleyer
Hey I just wanted to ask a quick formal question. The crate seems to have gathered quite a lot of contributors. However the paper only lists the submitting author as the sole author. Is this correct? There was for instance the project of `tqdm` where this was an issue, specifically regarding the attribution in the publication. During this time, JOSS did not assume an official stance on this but rather the contributors were forced to sort it out internally.

@Axect did you take any precautions/inform the other contributors etc.?
@jbytecode My comments from above are only hearsay. How does JOSS deal with this?

see `tqdm` Review process:
https://github.com/openjournals/joss-reviews/issues/1277

Edit: I will wait until this is sorted out before beginning with my review of the crate.
