### 2026-04-05 #10366 @Axect
@jonaspleyer Thank you for raising this. It's a fair question, and I appreciate you bringing it up before starting the review.

Regarding the comparison with tqdm: in that case, a reviewer identified a specific individual as a "significant contributor," which prompted the authorship discussion. In Peroxide's case, however, I have been the sole maintainer since the project's inception in 2018, and the commit history reflects this. Several git identities in the log (Axect, axect, "Tae Geun, Kim", "Tae Geun Kim", "Tae-Geun Kim") are all me, which accounts for over 90% of the total commits. The largest external contribution is 39 commits, and the rest are in the range of 1 to 13 commits.

I deeply value every single contribution to Peroxide. Since most external PRs were highly focused on specific bug fixes or localized enhancements rather than sustained core development, it did not occur to me to reach out to them regarding co-authorship for this paper.

However, I completely agree that their contributions should be properly recognized. I would greatly appreciate the editor's guidance on how JOSS typically handles such cases. Whether the recommendation is to include certain contributors as co-authors, explicitly list them by name in the Acknowledgements section, or take any other approach, I am fully open to it and will gladly follow JOSS's standard practices.

@jbytecode, could you please advise on this?
