### 2026-04-05 #10366 @jbytecode
@jonaspleyer - Thank you for your detailed review. It's a subjective question. Let me investigate the situation more deeply.

It looks like there are 19 contributors, and as @Axect states, much of the commit history is belong to the author. One of the leading contributor is [soumyasen1809](https://github.com/soumyasen1809) and they have much commits as well. They have some commits about adding new features, removing existing code, as well as adding bench-marking.

@Axect - In that point of review, I think, it's a nice experience to invite them if they wanted to contribute and having a place in the authors' list. That is a kind behavior in the academic publishing. How about opening a new issue and invite them as a authors?

@Axect - https://github.com/russellb23 and https://github.com/JSorngard and some others have relatively small number of commits, I can't measure how important are these commits, but again, it's a kind behavior to invite them if you believe their contributions add a value in your project.

RCaller, my first publication in JOSS, had many contributors and our editor had kindly asked me to ask them if they wanted to contribute in the paper, also. Here is the issue in which I managed the author list:

https://github.com/jbytecode/rcaller/issues/42

@Axect, would you be willing to make the same action?

@jonaspleyer - I hope my assist level satisfies your questions.
