### 2026-04-05 #10366 @jonaspleyer
Thanks for the quick reply. I believe that this is the right way. I also have a similar problem where I want to add a license to an existing crate `approx` which formally requires the confirmation of all authors.
https://github.com/brendanzab/approx/issues/94

I will start with my review of the software but will hold back on the paper itself until the other contributors have responded.
