### 2026-04-05 #10366 @Axect
Hi @jbytecode, thank you for the clear guidance and for sharing your positive experience with RCaller.

I completely agree with your suggestion. It is indeed a great practice to invite the contributors, and I am more than happy to take this approach. I will open an issue shortly in the Peroxide repository, tagging @soumyasen1809, @russellb23, @JSorngard, and all other contributors to ask if they would like to be included as co-authors.

I will leave the issue open for a reasonable amount of time to give them a chance to respond. Once we have their decisions, I will update the author list in paper.md accordingly, and ensure everyone is appropriately credited in the Acknowledgements section.

@jonaspleyer, thank you again for bringing this up. I will notify you and the editor here once the authorship is sorted out so we can proceed with the review.
