### 2026-04-05 Axect/Peroxide#94 @Axect
Invitation to Co-author the JOSS Paper
Hello everyone! 👋

I am currently in the process of submitting Peroxide to the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/). JOSS is a developer-friendly, open-access journal for research software packages.

During the review process, the JOSS editor and reviewer kindly suggested that I reach out to all contributors. I deeply appreciate the time and effort each of you has put into this project over the years, whether through adding new features, fixing bugs, or improving the codebase.

Because your contributions have been valuable to Peroxide, I would like to formally invite you to be listed as a co-author on the JOSS paper.

Since JOSS is designed to recognize software contributions, your previous commits to this repository already qualify you for co-authorship. You are welcome to review the current draft of the paper [here](https://github.com/Axect/Peroxide/blob/features/joss-paper/paper/paper.md) if you'd like, but no additional writing is required on your part.

### How to accept this invitation

If you would like to be included as a co-author, please reply to this issue **by April 26, 2026** with the following details required by JOSS:

- **Full Name**
- **Affiliation** (University, Company, or "Independent Researcher")
- **ORCID iD** (If you don't have one, you can register at https://orcid.org)

### If you decline or don't reply

If you prefer not to be a co-author, or if I don't hear back by April 26, 2026, that is completely fine. I will make sure to recognize your name in the Acknowledgements section of the paper.

For context, here is the JOSS review issue: https://github.com/openjournals/joss-reviews/issues/10366

Thank you all again for your contributions to Peroxide!

---

**Tagging contributors for visibility:**

@soumyasen1809 @russellb23 @JSorngard @schrieveslaach @GComitini @developing-human @bertini97 @mulimoen @samnaughtonb @adamnemecek @Hoff97 @relastle @jgrage @koute @tchamelot @era127 @tarolling @thettasch

(Nathan Eckert: if you see this, you are also invited. I was unable to find your GitHub account to tag you directly.)
