### 2026-04-08 Axect/Peroxide#94 @schrieveslaach
@Axect, thanks for reaching out. Sure, you can include me as author in the paper.

- Full Name: Marc Schreiber
- Affiliation: Independent Researcher (I think is appropriate in this case because work was done in my spare time)
- ORCID iD: https://orcid.org/0000-0003-2564-6126
