### 2026-04-08 Axect/Peroxide#94 @JSorngard
@Axect Thanks for the offer!

Full name: Johanna Sörngård
Affiliation: Independent Researcher (as I also did the work in my spare time)
ORCID: https://orcid.org/0000-0002-8660-9989
