### 2026-04-08 Axect/Peroxide#94 @tchamelot
Hello, thank you for reaching out. My contribution does not deserve any acknowledgement, it was not related to the mathematical aspects.

Good luck with the submission process and thank you for your work on this library
