### 2026-04-09 Axect/Peroxide#94 @JSorngard
Actually, my main contribution was in @Axect's `puruspe`. While `peroxide` re-exports that library I'm unsure if my contribution to `peroxide` proper warrants authorship. I would like to hear others thoughts on this.
