### 2026-04-09 Axect/Peroxide#94 @Axect
> Actually, my main contribution was in [@Axect](https://github.com/Axect)'s `puruspe`. While `peroxide` re-exports that library I'm unsure if my contribution to `peroxide` proper warrants authorship. I would like to hear others thoughts on this.

Peroxide is an open-source project built through the accumulation of each and every contribution from its contributors. While contributions may vary in size, every single one is valuable.

Unlike traditional academic journals, JOSS is a journal dedicated to open-source software, so I believe the bar for authorship is quite different, and your contributions more than clear it. After all, all of you are already listed as Peroxide contributors for a reason. 😉  (This isn't an official JOSS stance, but to my knowledge they don't impose restrictive authorship criteria for contributors either.)

On a personal note, your consistent contributions have been a great help to Peroxide. As you mentioned, your main work has been on puruspe, but since puruspe serves as the backbone of Peroxide's special function module, your contributions have had a significant, both direct and indirect, impact on the project as a whole. Thank you again for your contributions! 👍

So in my view, your authorship is more than warranted. I'd be glad to have you on the paper. 😄
