### 2026-04-09 Axect/Peroxide#94 @JSorngard
> So in my view, your authorship is more than warranted. I'd be glad to have you on the paper. 😄

Then I'll be happy to be there!
