### 2026-04-09 Axect/Peroxide#94 @GComitini
Hi @Axect, thanks for the invitation! Given what you wrote in this thread to me and others, I gladly accept to be co-author!

- Full Name: Giorgio Comitini
- Affiliation: Università degli Studi di Catania
- ORCID iD: https://orcid.org/0000-0002-1758-7878

My contributions were made in my spare time, but the code itself was developed as part of scientific research for my university, so I think the affiliation should be that one.
