### 2026-04-11 Axect/Peroxide#94 @olamide142
Hi @Axect ,
I’ve mainly been a supporter and user of the project rather than a direct contributor to the codebase, so I don’t think I meet the bar for co-authorship.
Happy to be acknowledged though
I’m a big fan of Peroxide and excited to see it published.
Thanks again for the great work.

Full Name: Victor Olowofeso
Affiliation: Independent Researcher
ORCID ID: https://orcid.org/0009-0002-8193-4405
