### 2026-04-15 Axect/Peroxide#94 @tarolling
Full Name: Benjamin Joens
Affiliation: Independent Researcher
ORCID: https://orcid.org/0009-0002-7922-0884

Appreciate you creating this library and bringing it all together :)
