### 2026-04-15 Axect/Peroxide#94 @soumyasen1809
@Axect Thank you very much for the offer.

Full Name: Soumya Sen
Affiliation: Independent Researcher (Same reason as in the replies above)
ORCID: https://orcid.org/0009-0008-0337-9964
