### 2026-04-19 b6273fe
Implement Dirichlet(α) distribution
