### 2026-04-19 ebf85c5
Implement Dirichlet(α) distribution
