+++ src/statistics/dist.rs
@@ -13,12 +13,14 @@
 //!     * Uniform
 //!     * Weighted Uniform
 //!     * Log Normal
-//! * There are two enums to represent probability distribution
+//! * There are three enums to represent probability distribution
 //!     * `OPDist<T>` : One parameter distribution (Bernoulli)
 //!     * `TPDist<T>` : Two parameter distribution (Uniform, Normal, Beta, Gamma)
+//!     * `MVDist<T>` : Multivariate distribution (Dirichlet)
 //!         * `T: PartialOrd + SampleUniform + Copy + Into<f64>`
 //! * There are some traits for pdf
-//!     * `RNG` trait - extract sample & calculate pdf
+//!     * `RNG` trait - extract sample & calculate pdf for 1D distributions
+//!     * `MVRNG` trait - extract sample & calculate pdf for multivariate distributions
 //!     * `Statistics` trait - already shown above
 //!
 //! ### `RNG` trait
@@ -239,6 +241,55 @@
 //!     * Mean: $e^{\mu + \frac{\sigma^2}{2}}$
 //!     * Var: $(e^{\sigma^2} - 1)e^{2\mu + \sigma^2}$
 //! * To generate log-normal random samples, Peroxide uses the `rand_distr::LogNormal` distribution from the `rand_distr` crate.
+//! ### `MVRNG` trait
+//!
+//! * `MVRNG` trait is composed of four fields
+//!     * `sample`: Extract samples
+//!     * `sample_with_rng`: Extract samples with specific rng
+//!     * `pdf` : Calculate pdf value at specific point
+//!     * `ln_pdf` : Calculate log-pdf value at specific point
+//!     ```no_run
+//!     use rand::Rng;
+//!     pub trait MVRNG {
+//!         /// Extract samples of multivariate distributions
+//!         fn sample(&self, n: usize) -> Vec<Vec<f64>>;
+//!
+//!         /// Extract samples of distributions with specific rng
+//!         fn sample_with_rng<R: Rng + Clone>(&self, rng: &mut R, n: usize) -> Vec<Vec<f64>>;
+//!
+//!         /// Probability Density Function
+//!         fn pdf(&self, x: &[f64]) -> f64;
+//!
+//!         /// Log Probability Density Function
+//!         fn ln_pdf(&self, x: &[f64]) -> f64;
+//!     }
+//!     ```
+//!
+//! ### Dirichlet Distribution
+//!
+//! * Definition
+//!     $$ \text{Dir}(\mathbf{x} | \boldsymbol{\alpha}) = \frac{1}{\text{B}(\boldsymbol{\alpha})} \prod_{i=1}^K x_i^{\alpha_i - 1} $$
+//!     where $\text{B}(\boldsymbol{\alpha}) = \frac{\prod_{i=1}^K \Gamma(\alpha_i)}{\Gamma(\sum_{i=1}^K \alpha_i)}$
+//! * Representative value
+//!     * Mean: $\frac{\alpha_i}{\alpha_0}$
+//!     * Var : $\frac{\alpha_i(\alpha_0 - \alpha_i)}{\alpha_0^2(\alpha_0 + 1)}$
+//! * To generate Dirichlet random samples, Peroxide generates $K$ independent Gamma samples and normalizes them.
+//! * **Caution**: `MVDist` utilizes the existing `Statistics` trait but outputs vectors and matrices.
+//!
+//!     ```rust
+//!     use peroxide::fuga::*;
+//!
+//!     fn main() {
+//!         let mut rng = smallrng_from_seed(42);
+//!         let a = Dirichlet(vec![1.0, 2.0, 3.0]);     // Dir(x | 1.0, 2.0, 3.0)
[…]
