### 2026-04-19 15423c6
Implement `Print`
