### 2026-04-19 bd28135
Implement `Print`
