+++ src/util/print.rs
@@ -320,265 +320,6 @@ impl Printable for &Vec<f64> {
     }
 }
 
-impl Printable for Vec<Vec<usize>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<u8>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<u16>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<u32>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<u64>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<isize>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<i8>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<i16>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<i32>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<i64>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<char>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<&str>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<String>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for Vec<Vec<bool>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<usize>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<u8>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<u16>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<u32>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<u64>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<isize>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<i8>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<i16>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<i32>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<i64>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<char>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<&str>> {
-    fn print(&self) {
-        println!("{:?}", self);
-    }
-}
-
-impl Printable for &Vec<Vec<String>> {
[…]
