### 2026-04-29 da19038
PAPER: Add 6 co-authors and update Acknowledgements
- Add Comitini, Grage, Joens, Schreiber, Sen, Sörngård as co-authors
  (alphabetical by last name) following contributor invitation in #94
- Update affiliations: add Università degli Studi di Catania (3) and
  Independent Researcher (4)
- Mark Tae-Geun Kim as corresponding author
- Reword AI usage disclosure for multi-author context
- Compress Acknowledgements; explicitly thank Victor Olowofeso
- Sync CITATION.cff authors

Word count: 1747 / 1750.
