### 2026-04-29 #10366 @Axect
@jbytecode Thank you for your patience. Here is the status update following the April 26 deadline for contributor co-author invitations [(Peroxide issue 94)](https://github.com/Axect/Peroxide/issues/94).

- New co-authors (6, listed alphabetically by last name after the corresponding author): Giorgio Comitini, Jonas Grage, Benjamin Joens, Marc Schreiber, Soumya Sen, and Johanna Sörngård. Their affiliations and ORCIDs are reflected in the updated paper/paper.md and CITATION.cff.

- Other responses: Three contributors replied but declined co-authorship; one of them is acknowledged by name in the paper at their request, and the others are recognized through the repository's contributor history. The remaining tagged contributors did not respond by the deadline and are likewise recognized through the contributor history, as stated in the original invitation.

- Updates pushed to the features/joss-paper branch.

Please let me know if any further changes are needed. Thanks again for your guidance throughout the review.
