### 2026-05-06 Axect/Peroxide#97 @jonaspleyer
[JOSS-Review] Fix build of docs on docs.rs
The docs are not building on docs.rs. Since crates.io and docs.rs are the official platform for releasing documentation of crates and possibly more persistent than Github's pages, it would make sense to try and resolve this build problem.

The issue seems to stem from a feature which is active in the release but not supported anymore by the latest nightly version of the compiler. Try to install the latest nightly version and build with this version.

```
rustup install nightly
cargo +nightly doc --no-deps
```

https://github.com/openjournals/joss-reviews/issues/10366
