### 2026-05-06 Axect/Peroxide#98 @jonaspleyer
[JOSS-Review] Document features more carefully
I have run the test suite and examples with for all individual features and the combination of 2 features. These are the results:

```
✅ = Passed
❌ = Failed
⏩ = Skipped (due to previous Fail)
1st col: debug mode
2nd col: release mode
```

When trying to test with `O3` it seems that I am missing some dependencies but there are also import errors in `tests/o3/lapack.rs` and others.

Building with the feature `nc` or `netcdf` fails due to missing HDF5 headers.

Building with the feature `plot` works fine, but testing results in errors for various tests:
```
test src/numerical/spline.rs - numerical::spline (line 97) ... FAILED
test src/numerical/ode.rs - numerical::ode (line 39) ... FAILED
test src/numerical/spline.rs - numerical::spline::CubicBSplineBases (line 974) ... FAILED
test src/util/plot.rs - util::plot (line 16) ... FAILED
```

Overall I would expect either links to the necessary dependencies/crates which are activated by the features. And in the case of any necessary python dependencies, I would expect a description of how to prepare an environment such that the test suite can be run.

This information should be visible from the documentation. Furthermore since docs.rs is currently not building the latest version, there is no complete feature list for the most recent version.

See: https://github.com/Axect/Peroxide/issues/97

List of full tested configs
```
# Testing Examples
✅✅ ad
✅✅ dataframe
✅✅ gradient
✅✅ lorenz_gl4
✅✅ norm
✅✅ ode_test_rkf78
✅✅ root_test
✅✅ area_under_spline
✅✅ det
✅✅ hessian
✅✅ lorenz_rk4
✅✅ ode_env
✅✅ optim
✅✅ solve
✅✅ broyden_test
✅✅ df_csv
✅✅ higher_order_ad
✅✅ lorenz_rkf45
✅✅ ode_test_gl4
✅✅ poly
✅✅ spread
✅✅ bspline_calculus
✅✅ df_nc
✅✅ integrate
✅✅ lorenz_rkf78
✅✅ ode_test_orbit
✅✅ pseudo_inv
✅✅ svd
✅✅ b_spline_test
✅✅ df_parquet
✅✅ jet_ad
✅✅ lorenz_tsit45
✅✅ ode_test_rk4
✅✅ real_trait_test
✅✅ thomas
✅✅ chebyshev
✅✅ dist
✅✅ lorenz_dp45
✅✅ matmul
✅✅ ode_test_rkf45
✅✅ root_macro_test

# Feature Combinations: 1
✅✅ default
❌❌ O3
✅✅ arrow
✅✅ blas
✅✅ complex
✅✅ csv
✅✅ indexmap
✅✅ json
✅✅ lapack
❌❌ nc
❌❌ netcdf
✅✅ num-complex
✅✅ parallel
✅✅ parquet
❌❌ plot
✅✅ pyo3
✅✅ rayon
✅✅ rkyv
✅✅ serde

# Feature Combinations: 2
⏩   default,O3
✅✅ default,arrow
✅✅ default,blas
✅✅ default,complex
✅✅ default,csv
✅✅ default,indexmap
✅✅ default,json
✅✅ default,lapack
⏩   default,nc
⏩   default,netcdf
✅✅ default,num-complex
✅✅ default,parallel
✅✅ default,parquet
⏩   default,plot
✅✅ default,pyo3
✅✅ default,rayon
✅✅ default,rkyv
✅✅ default,serde
⏩   O3,arrow
⏩   O3,blas
⏩   O3,complex
⏩   O3,csv
⏩   O3,indexmap
⏩   O3,json
⏩   O3,lapack
⏩   O3,nc
⏩   O3,netcdf
⏩   O3,num-complex
⏩   O3,parallel
⏩   O3,parquet
⏩   O3,plot
⏩   O3,pyo3
⏩   O3,rayon
⏩   O3,rkyv
⏩   O3,serde
✅✅ arrow,blas
✅✅ arrow,complex
✅✅ arrow,csv
✅✅ arrow,indexmap
✅✅ arrow,json
✅✅ arrow,lapack
⏩   arrow,nc
⏩   arrow,netcdf
✅✅ arrow,num-complex
✅✅ arrow,parallel
✅✅ arrow,parquet
⏩   arrow,plot
✅✅ arrow,pyo3
✅✅ arrow,rayon
✅✅ arrow,rkyv
✅✅ arrow,serde
✅✅ blas,complex
✅✅ blas,csv
✅✅ blas,indexmap
✅✅ blas,json
✅✅ blas,lapack
⏩   blas,nc
⏩   blas,netcdf
✅✅ blas,num-complex
✅✅ blas,parallel
✅✅ blas,parquet
⏩   blas,plot
✅✅ blas,pyo3
✅✅ blas,rayon
✅✅ blas,rkyv
✅✅ blas,serde
✅✅ complex,csv
✅✅ complex,indexmap
✅✅ complex,json
✅✅ complex,lapack
⏩   complex,nc
⏩   complex,netcdf
✅✅ complex,num-complex
✅✅ complex,parallel
✅✅ complex,parquet
⏩   complex,plot
✅✅ complex,pyo3
✅✅ complex,rayon
✅✅ complex,rkyv
✅✅ complex,serde
✅✅ csv,indexmap
✅✅ csv,json
✅✅ csv,lapack
⏩   csv,nc
⏩   csv,netcdf
✅✅ csv,num-complex
✅✅ csv,parallel
✅✅ csv,parquet
⏩   csv,plot
✅✅ csv,pyo3
✅✅ csv,rayon
✅✅ csv,rkyv
✅✅ csv,serde
✅✅ indexmap,json
✅✅ indexmap,lapack
⏩   indexmap,nc
⏩   indexmap,netcdf
✅✅ indexmap,num-complex
✅✅ indexmap,parallel
✅✅ indexmap,parquet
⏩   indexmap,plot
✅✅ indexmap,pyo3
✅✅ indexmap,rayon
✅✅ indexmap,rkyv
✅✅ indexmap,serde
✅✅ json,lapack
⏩   json,nc
⏩   json,netcdf
✅✅ json,num-complex
✅✅ json,parallel
✅✅ json,parquet
⏩   json,plot
✅✅ json,pyo3
✅✅ json,rayon
✅✅ json,rkyv
✅✅ json,serde
⏩   lapack,nc
⏩   lapack,netcdf
✅✅ lapack,num-complex
✅✅ lapack,parallel
✅✅ lapack,parquet
⏩   lapack,plot
✅✅ lapack,pyo3
✅✅ lapack,rayon
✅✅ lapack,rkyv
✅✅ lapack,serde
⏩   nc,netcdf
⏩   nc,num-complex
⏩   nc,parallel
⏩   nc,parquet
⏩   nc,plot
⏩   nc,pyo3
⏩   nc,rayon
⏩   nc,rkyv
⏩   nc,serde
⏩   netcdf,num-complex
⏩   netcdf,parallel
⏩   netcdf,parquet
⏩   netcdf,plot
⏩   netcdf,pyo3
⏩   netcdf,rayon
⏩   netcdf,rkyv
⏩   netcdf,serde
✅✅ num-complex,parallel
✅✅ num-complex,parquet
⏩   num-complex,plot
✅✅ num-complex,pyo3
✅✅ num-complex,rayon
✅✅ num-complex,rkyv
✅✅ num-complex,serde
✅✅ parallel,parquet
⏩   parallel,plot
✅✅ parallel,pyo3
✅✅ parallel,rayon
✅✅ parallel,rkyv
✅✅ parallel,serde
⏩   parquet,plot
✅✅ parquet,pyo3
✅✅ parquet,rayon
✅✅ parquet,rkyv
✅✅ parquet,serde
⏩   plot,pyo3
⏩   plot,rayon
⏩   plot,rkyv
⏩   plot,serde
✅✅ pyo3,rayon
✅✅ pyo3,rkyv
✅✅ pyo3,serde
✅✅ rayon,rkyv
✅✅ rayon,serde
✅✅ rkyv,serde
```

https://github.com/openjournals/joss-reviews/issues/10366#issuecomment-4187914529
