### 2026-05-06 Axect/Peroxide#98 @jonaspleyer
Addendum: I have just noticed that some information is actually provided in the `lib.rs` file. However I could not immediately find it. Consider merging the `Pre-requisite` section with the section `Available Features`.
