### 2026-05-06 Axect/Peroxide#99 @jonaspleyer
[JOSS-Review] Suggestion: Restructure README.md
It seems that the landing page of the documentation build with `cargo doc` and the contents of the `README.md` are very close to each other, or almost identical.

I would suggest to focus your efforts and to provide the some of the information only in the technical documentation build by `cargo doc`. My suggestion is to remove the following subsections from the `README.md` file:

```
Latest README version
Useful tips for features
Module Structure
Documentation
Examples
Release Info (maybe)
Contributing Guide
LICENSE
TODO
```

I furthermore suggest to shorten the list of bullet points to contain more than one variant per category. i.e. changes like so:

```
# PREVIOUS
- Probability Distributions
  - Bernoulli
  - Uniform
  - ...

# NEW
- Probability Distributions (Bernoulli, Uniform, ...)
```

https://github.com/openjournals/joss-reviews/issues/10366#issuecomment-4187914529
