### 2026-05-06 Axect/Peroxide#100 @jonaspleyer
[JOSS-Review] Paper suggestions
Here is a unstructured list of suggestions to improve the paper:

- [x] Add LAPACK citation (line 29)
- [x] Add `ndarray` citation (Table 1; Description)
- [x] Briefly discuss possible known downsides of the matrix memory layout as `Vec<f64>` as opposed to other approaches. I believe there were some blog posts when `ndarray` was first created about this topic (maybe rather `nalgebra` I am not completely certain)
- [x] Include enzyme in the discussion. How does this compare to the AD functionality of peroxide? https://github.com/EnzymeAD/Enzyme https://docs.rs/enzyme/latest/enzyme/ https://www.youtube.com/watch?v=FR4fwGsBYPI

I will continue to add more points to this issue as I continue my review.

https://github.com/openjournals/joss-reviews/issues/10366#issuecomment-4187914529
