### 2026-05-09 70d3533
DOCS: Fix docs.rs build (JOSS review #97)
Removed `#![cfg_attr(docsrs, feature(doc_auto_cfg))]` because the
`doc_auto_cfg` nightly feature was removed in Rust 1.92 and merged into
`doc_cfg` (RFC 3631). Auto-cfg badges are now the rustdoc default, so
no feature gate is required. Dropped the now-unused `--cfg docsrs`
rustdoc arg accordingly.

Also enumerated the buildable features in `[package.metadata.docs.rs]`
so feature-gated APIs (csv/parquet/blas/lapack/etc.) appear in the
published docs. Excludes `O3`, `nc`/`netcdf`, and `plot` which require
system dependencies unavailable in the docs.rs s
[…]
