### 2026-05-09 9141070
TEST: Fix tests/o3 and tests/blas_test compile errors (JOSS review #98)
Two latent test breakages exposed by the matrix testing in JOSS review
issue #98:

* `tests/o3/lapack.rs` referenced `UPLO::Upper` / `UPLO::Lower` through
  `peroxide::fuga::*`, but `fuga` only re-exported the variants, not
  the enum type. Re-export `UPLO` itself so the documented `cholesky`
  /`lapack_dpotrf` API is usable from `fuga`.

* `tests/blas_test.rs` was missing `#[macro_use]` on its
  `extern crate peroxide` declaration (`c!` macro unresolved) and the
  `dgemv_test` expected `Matrix * Vec<f64>` to return a `Mat
[…]
