### 2026-05-09 cd73e1b
DOCS: Mark plot doctests as no_run (JOSS review #98)
The four doctests reported failing under `--features plot` in the
review (src/util/plot.rs L16, src/numerical/spline.rs L97 + L974,
src/numerical/ode.rs L39) all end with `.savefig(...)`, which spawns a
Python process via pyo3 and writes to `example_data/`. That makes them
unrunnable in any environment without matplotlib + a writable target
path -- including docs.rs, the JOSS reviewer's host, and most CI.

Switch the four affected fences to `no_run` so they continue to verify
that the example compiles without requiring the matplotlib runtime
[…]
