### 2026-05-09 fc1c41a
DOCS: Expand feature documentation in README (JOSS review #98)
The previous Pre-requisite and Available Features sections were
sparse: they named OpenBLAS / matplotlib / netcdf without explaining
how to install them, and the feature list was incomplete relative to
`Cargo.toml`. Reviewers reported broken builds and could not tell
which features needed which dependencies.

* Pre-requisite: replace the three-bullet note with per-feature
  sections for `O3` / `plot` / `nc`, each with platform-specific
  install commands (Debian, Fedora, Arch, macOS) and a pointer to the
  underlying Rust crate. Ad
[…]
