### 2026-05-09 95c42c5
BUILD: Add O3-* convenience features for BLAS/LAPACK link backend (JOSS review #98)
The bare `O3` feature only activates the `blas` and `lapack` FFI
wrappers, which leaves the `dgemv_` / `dpotrf_` / ... symbols
unresolved at link time. Reviewers and CI both hit `undefined symbol`
errors when running `cargo test --features O3` even with OpenBLAS
installed system-wide, because nothing pulls in `blas-src` /
`lapack-src`.

Add four convenience features mirroring the standard practice in
`ndarray-linalg`:

* `O3-openblas`   - OpenBLAS (Linux / Windows / macOS via Homebrew)
* `O3-accelerate` - Apple
[…]
