### 2026-05-09 d4e3c9c
DOCS: Document O3 backend selection and HDF5 1.x constraint (JOSS review #98)
Follow up on the feature documentation rewrite by replacing the
"install OpenBLAS" hand-wave with a full explanation of how the BLAS
link backend is chosen:

* Pre-requisite: introduce the new `O3-openblas` / `O3-accelerate` /
  `O3-mkl` / `O3-netlib` convenience flags as the recommended path
  and explain that bare `O3` users must add `blas-src` / `lapack-src`
  to their downstream Cargo.toml. Note that `O3-accelerate` and
  `O3-mkl` ship their own backend.

* Pre-requisite (`nc` / `netcdf`): document that the pinne
[…]
