### 2026-05-09 093f4b1
CI: Test optional features in separate jobs (JOSS review #98)
The previous workflow only ran `cargo build` and `cargo test` with
default features, so all four optional-feature regressions surfaced
in the JOSS review (`O3` UPLO breakage, `tests/blas_test` macro
import, `plot` doctests, the `O3` link backend gap) were latent for
release after release. Nothing in CI exercised them.

Split the single job into four:

* `default`     - keeps the original cargo build + cargo test path.
* `o3`          - installs OpenBLAS + LAPACK + gfortran and runs the
                  test suite under `O3-openblas
[…]
