### 2026-05-12 ada3d19
LINT: Fix clippy errors, add informational CI job (JOSS ariostas)
`cargo clippy --lib --no-deps` was emitting two deny-level errors that
blocked introducing clippy into CI:

* `src/numerical/integral.rs:451` used the literal `1.4142135623730951`
  in a `tol / sqrt(2)` expression. Switch to `std::f64::consts::SQRT_2`
  so the constant is named.

* `src/structure/dataframe.rs:1182` defined an inherent
  `Scalar::to_string(self)` that shadows the `Display`-derived
  `to_string(&self)`. The two intentionally produce different output -
  the inherent variant returns the bare value while `Display` a
[…]
