### 2026-05-12 db3e340
DOCS: Expand CONTRIBUTING.md (JOSS ariostas)
The previous CONTRIBUTING.md was three Gitflow bullets and nothing
else - no bug-report guidance, no test/lint workflow, no link to the
Code of Conduct. ariostas asked for the file to be cleaned up and to
invite issue reports.

Rewrite the file to cover:

* How to file bug reports and feature requests, including the minimal
  information that helps triage (snippet, Peroxide version, Cargo
  features, rustc version, OS, backend libraries).
* Where to ask non-bug questions.
* The Gitflow contribution rules, with concrete `cargo test --features
  ...` 
[…]
