### 2026-05-12 3ec5e46
LINT: Apply auto-fixable clippy suggestions (JOSS ariostas)
Run `cargo clippy --fix --lib --no-deps` across the four CI feature
groups (default, O3-openblas, pure-Rust, plot) and accept every
machine-applicable suggestion. The fixes are mechanical: collapsed
needless `clone()` / `into_iter()` calls, removed unneeded `return`
statements, replaced loop-by-index with iterator forms, swapped
redundant closures for direct function references, and similar
restyles that do not change observable behaviour.

Lint counts on the lib (full suite still has plot doctests, examples,
etc.):

  Feature set    
[…]
