### 2026-05-13 bf128a0
LINT: Resolve cosmetic clippy warnings (JOSS ariostas)
Follow up on the auto-fix pass (3ec5e46) with the manual fixes for
warnings rustfix cannot apply on its own. All edits are
behaviour-preserving:

* `assertions_on_constants` (eigen.rs): replace `assert!(false, msg)`
  with `panic!(msg)` (the lint's recommended form).
* `empty_line_after_doc_comments` (root.rs): drop the blank line
  separating a doc block from the item it documents.
* `while_let_loop` (integral.rs): collapse `loop { match Vec::pop() }`
  to `while let Some(...) = S.pop()`.
* `len_without_is_empty` (dataframe.rs): add the m
[…]
