+++ src/statistics/dist.rs
@@ -48,7 +48,7 @@
 //! ### Bernoulli Distribution
 //!
 //! * Definition
-//!     $$ \text{Bern}(x | \mu) = \mu^x (1-\mu)^{1-x} $$
+//!   $$ \text{Bern}(x | \mu) = \mu^x (1-\mu)^{1-x} $$
 //! * Representative value
 //!     * Mean: $\mu$
 //!     * Var : $\mu(1 - \mu)$
@@ -75,7 +75,7 @@
 //! ### Uniform Distribution
 //!
 //! * Definition
-//!     $$\text{Unif}(x | a, b) = \begin{cases} \frac{1}{b - a} & x \in \[a,b\]\\\ 0 & \text{otherwise} \end{cases}$$
+//!   $$\text{Unif}(x | a, b) = \begin{cases} \frac{1}{b - a} & x \in \[a,b\]\\\ 0 & \text{otherwise} \end{cases}$$
 //! * Representative value
 //!     * Mean: $\frac{a + b}{2}$
 //!     * Var : $\frac{1}{12}(b-a)^2$
@@ -99,7 +99,7 @@
 //! ### Normal Distribution
 //!
 //! * Definition
-//!     $$\mathcal{N}(x | \mu, \sigma^2) = \frac{1}{\sqrt{2\pi \sigma^2}} \exp{\left( - \frac{(x - \mu)^2}{2\sigma^2}\right)}$$
+//!   $$\mathcal{N}(x | \mu, \sigma^2) = \frac{1}{\sqrt{2\pi \sigma^2}} \exp{\left( - \frac{(x - \mu)^2}{2\sigma^2}\right)}$$
 //! * Representative value
 //!     * Mean: $\mu$
 //!     * Var: $\sigma^2$
@@ -126,8 +126,8 @@
 //! ### Beta Distribution
 //!
 //! * Definition
-//!     $$\text{Beta}(x | \alpha, \beta) = \frac{1}{\text{B}(\alpha, \beta)} x^{\alpha-1} (1-x)^{\beta-1}$$
-//!     where $\text{B}(\alpha, \beta) = \frac{\Gamma(\alpha)\Gamma(\beta)}{\Gamma(\alpha+\beta)}$ is the Beta function.
+//!   $$\text{Beta}(x | \alpha, \beta) = \frac{1}{\text{B}(\alpha, \beta)} x^{\alpha-1} (1-x)^{\beta-1}$$
+//!   where $\text{B}(\alpha, \beta) = \frac{\Gamma(\alpha)\Gamma(\beta)}{\Gamma(\alpha+\beta)}$ is the Beta function.
 //! * Representative value
 //!     * Mean: $\frac{\alpha}{\alpha+\beta}$
 //!     * Var: $\frac{\alpha\beta}{(\alpha+\beta)^2(\alpha+\beta+1)}$
@@ -149,8 +149,8 @@
 //! ### Gamma Distribution
 //!
 //! * Definition
-//!     $$\text{Gamma}(x | \alpha, \beta) = \frac{\beta^\alpha}{\Gamma(\alpha)} x^{\alpha-1} e^{-\beta x}$$
-//!     where $\Gamma(\alpha) = \int_0^\infty x^{\alpha-1} e^{-x} dx$ is the Gamma function.
+//!   $$\text{Gamma}(x | \alpha, \beta) = \frac{\beta^\alpha}{\Gamma(\alpha)} x^{\alpha-1} e^{-\beta x}$$
+//!   where $\Gamma(\alpha) = \int_0^\infty x^{\alpha-1} e^{-x} dx$ is the Gamma function.
 //! * Representative value
 //!     * Mean: $\frac{\alpha}{\beta}$
[…]
