+++ src/structure/matrix.rs
@@ -445,7 +445,7 @@
 //! * Peroxide uses LU decomposition (via GECP) to obtain inverse matrix.
 //! * It needs two sub functions - `inv_l`, `inv_u`
 //!     * For inverse of `L, U`, I use block partitioning. For example, for lower triangular matrix :
-//!     $$ \begin{aligned} L &= \begin{pmatrix} L_1 & \mathbf{0} \\\ L_2 & L_3 \end{pmatrix} \\\ L^{-1} &= \begin{pmatrix} L_1^{-1} & \mathbf{0} \\\ -L_3^{-1}L_2 L_1^{-1} & L_3^{-1} \end{pmatrix} \end{aligned} $$
+//!       $$ \begin{aligned} L &= \begin{pmatrix} L_1 & \mathbf{0} \\\ L_2 & L_3 \end{pmatrix} \\\ L^{-1} &= \begin{pmatrix} L_1^{-1} & \mathbf{0} \\\ -L_3^{-1}L_2 L_1^{-1} & L_3^{-1} \end{pmatrix} \end{aligned} $$
 //!     ```rust
 //!     #[macro_use]
 //!     extern crate peroxide;
@@ -467,12 +467,10 @@
 //! ```rust
 //! use peroxide::fuga::*;
 //!
-//! fn main() {
-//!     let a = ml_matrix("1 2;3 4");
-//!     let pqlu = a.lu();  // Memoization of LU
-//!     pqlu.det().print(); // Same as a.det() but do not need an additional LU
-//!     pqlu.inv().print(); // Same as a.inv() but do not need an additional LU
-//! }
+//! let a = ml_matrix("1 2;3 4");
+//! let pqlu = a.lu();  // Memoization of LU
+//! pqlu.det().print(); // Same as a.det() but do not need an additional LU
+//! pqlu.inv().print(); // Same as a.inv() but do not need an additional LU
 //! ```
 //!
 //! ## QR Decomposition (`O3` feature only)
