### 2026-05-13 13cb279
LINT: Resolve len_zero / missing_safety_doc / ptr_arg warnings (JOSS ariostas)
Phase 1 of the residual clippy cleanup. Three lint categories with
zero behavioural risk:

* `len_zero` (2 sites in dataframe.rs): replace `self.len() > 0` with
  `!self.is_empty()`, using the `Series::is_empty` shim added in
  bf128a0.

* `missing_safety_doc` (9 sites across mutable.rs / pointer.rs /
  low_level.rs): add proper `# Safety` sections to every `unsafe fn`
  signature in the public traits and helpers, spelling out the
  preconditions on indices, pointer validity, aliasing, and lifetime.

* `ptr_arg` (19
[…]
