### 2026-05-13 2f33cb2
TEST: Add clippy_verify determinism example for the Phase 2 refactor (JOSS ariostas)
Some `cargo clippy --fix` candidates - notably `needless_range_loop`
suggestions in algorithm code (matrix, ODE step, polynomial,
quadrature, Lanczos approximation, distributions) - cannot be applied
mechanically with the existing test suite as the only safety net. A
benign-looking rewrite that re-orders or skips an iteration would
produce different f64 results without necessarily failing any
hand-written `assert!`.

Add `examples/clippy_verify.rs` as an explicit determinism oracle.
The program exercises every
[…]
