+++ examples/clippy_verify.rs
@@ -1,0 +1,270 @@
+//! Determinism oracle for the Phase 2 `needless_range_loop` refactor
+//! (JOSS review feedback from ariostas).
+//!
+//! Each `cargo clippy --fix` suggestion for an index-based loop is
+//! semantically equivalent to the original only if it preserves bitwise
+//! output. This example exercises every public entry point that
+//! transitively invokes one of the 30 affected loops, with seeded
+//! random inputs, and prints a hash + a small spot-check value for
+//! each result.
+//!
+//! Workflow:
+//!
+//! ```ignore
+//! # snapshot baseline before touching algorithm code
+//! cargo run --example clippy_verify > /tmp/clippy_verify.golden
+//! # apply needless_range_loop refactors, then verify equivalence
+//! cargo run --example clippy_verify > /tmp/clippy_verify.after
+//! diff /tmp/clippy_verify.golden /tmp/clippy_verify.after  # must be empty
+//! ```
+
+use std::hash::{DefaultHasher, Hash, Hasher};
+
+use peroxide::fuga::*;
+use peroxide::numerical::integral::kronrod_quadrature;
+use peroxide::numerical::interp::chebyshev_nodes;
+use peroxide::special::lanczos::ln_gamma_approx;
+use peroxide::structure::sparse::SPMatrix;
+use peroxide::util::useful::{max, min};
+use rand::{rngs::SmallRng, SeedableRng};
+
+fn hash_f64s(v: &[f64]) -> String {
+    let mut h = DefaultHasher::new();
+    v.len().hash(&mut h);
+    for x in v {
+        x.to_bits().hash(&mut h);
+    }
+    format!("{:016x}", h.finish())
+}
+
+fn hash_matrix(m: &Matrix) -> String {
+    let mut h = DefaultHasher::new();
+    m.row.hash(&mut h);
+    m.col.hash(&mut h);
+    // m.shape is an enum without Hash; use Debug as a stable surrogate
+    format!("{:?}", m.shape).hash(&mut h);
+    for x in &m.data {
+        x.to_bits().hash(&mut h);
+    }
+    format!("{:016x}", h.finish())
+}
+
+fn print_kv<T: std::fmt::Display>(label: &str, value: T) {
+    println!("{:48} {}", label, value);
+}
+
+fn check_integral() {
+    // kronrod_quadrature drives `compute_gauss_kronrod_sums_stored`,
+    // which is the function holding the flagged loop. Call it directly
+    // at every supported order so the index-based body fires for each
+    // node count.
+    let f = |x: f64| x.cos() + 0.25 * x * x;
+    for &n in &[15_usize, 21, 31, 41, 51, 61] {
+        let r: f64 = kronrod_quadrature(f, n, (0.0, 1.7));
+        print_kv(
+            &format!("kronrod_quadrature n={}", n),
+            format!("{r:.17e}"),
+        );
+        // also exercise with a different smooth integrand to widen coverage
+        let r2: f64 = kronrod_quadrature(|x: f64| (-x * x).exp(), n, (-2.5, 1.5));
+        print_kv(
+            &format!("kronrod_quadrature gauss n={}", n),
+            format!("{r2:.17e}"),
+        );
+    }
+}
+
+fn check_chebyshev_lagrange() {
+    // chebyshev_nodes carries one of the affected loops directly.
+    for &n in &[4_usize, 8, 16, 32] {
[…]
