### 2026-05-13 4514acd
LINT: Resolve all remaining needless_range_loop warnings (JOSS ariostas)
Apply the `needless_range_loop` cleanup to every site flagged by
clippy across every CI feature configuration. Each rewrite was
sanity-checked against the bit-exact baseline produced by
`examples/clippy_verify` (commit 2f33cb2); every test in the four CI
feature jobs still passes (252 / 253 / 253 / 286 for default /
O3-openblas / plot / pure-Rust).

By file:

* `src/numerical/integral.rs` - kronrod node enumeration replaced by
  a slice iterator.
* `src/numerical/interp.rs` - chebyshev_nodes uses `iter_mut`.
* `src/numeri
[…]
