### 2026-05-13 5562bfb
LINT: Clear clippy --all-targets across tests and examples (JOSS ariostas)
The previous Phase 1 / Phase 2 cleanup only drove `cargo clippy --lib`
to zero warnings. Running `cargo clippy --all-targets` still surfaced
28 clippy issues (three of which - `approx_constant` - are `#[deny]`
by default and would block CI as soon as the job is widened to all
targets) plus a dozen rustc-level cosmetic warnings in the test and
example crates.

Fixes by category:

* `println_empty_string` (11): replace `println!("")` with
  `println!()` in `examples/dataframe.rs` and `examples/dist.rs`.
* `unit_arg` (5): 
[…]
