### 2026-05-13 c94a0c8
CI: Promote clippy job to --all-targets and make it blocking (JOSS ariostas)
The informational `clippy` job introduced in 093f4b1 only ran
`cargo clippy --lib --no-deps`, which left tests / examples / benches
uninstrumented. Several latent issues persisted there even after the
library itself was clean (now cleared in 5562bfb).

* Switch the command to `cargo clippy --all-targets --no-deps
  --features O3-openblas` so the full target graph is linted.
* Drop `continue-on-error: true` so a regression in any new
  contribution actually fails the pipeline. The lib is at zero
  warnings under defaul
[…]
