### 2026-05-15 9723fde
FIX: Allocate row_ics with the correct nnz length in SPMatrix::new
`SPMatrix::new(row, col, nnz)` initialised `row_ics` with
`vec![0usize, nnz]` (a literal two-element vector `[0, nnz]`) instead
of `vec![0usize; nnz]` (an `nnz`-length zeroed vector). Anything that
allocated a fresh `SPMatrix` for in-place writes - most importantly
`SPMatrix::transpose` - then panicked with
`index out of bounds: the len is 2 but the index is K` on any matrix
with more than two non-zeros.

The bug surfaced while wiring the sparse-transpose path into
`examples/clippy_verify` during the Phase 2 clippy refactor; th
[…]
