### 2026-05-15 8901470
DOCS: Point documentation URL to docs.rs (JOSS review ariostas #1 / jonaspleyer)
Both reviewers asked for docs.rs to become the canonical documentation
entry point instead of the manually-updated `Peroxide_Doc` GitHub
Pages site. The docs.rs build itself was unblocked back in 70d3533;
this follow-up moves the link targets so users land there by default.

* `Cargo.toml`: `documentation` field switched from
  `https://axect.github.io/Peroxide_Doc` to `https://docs.rs/peroxide`.
* `README.md`: the two `[![On docs.rs](docs.rs badge)](...)` shields
  (top banner and the Documentation section) now a
[…]
