+++ Cargo.toml
@@ -8,7 +8,7 @@ repository = "https://github.com/Axect/Peroxide"
 license = "MIT OR Apache-2.0"
 categories = ["science"]
 readme = "README.md"
-documentation = "https://axect.github.io/Peroxide_Doc"
+documentation = "https://docs.rs/peroxide"
 keywords = ["Numeric", "Science", "Dataframe", "Plot", "LinearAlgebra"]
 exclude = [
   "example_data/",
