### 2026-05-15 63c881c
DOCS: Add Statement of need and examples links to crate landing page (JOSS ariostas #2)
The docs.rs landing page used to open with a single line ("`peroxide`
is comprehensive numerical library for Rust.") before jumping straight
to a flat component listing. Reviewer ariostas flagged this against
the JOSS "Statement of need" checklist item and asked for links to
the in-repo `examples/` directory and the companion `Peroxide_Gallery`.

Replace the one-liner with three short paragraphs adapted from the
JOSS paper's Statement of need:

* what scientific workflows look like in practice and how SciPy
[…]
