### 2026-05-15 86cc03d
DOCS: Restructure README, add quickstart and source layout (JOSS #99)
Address jonaspleyer's review item #99 (README / cargo doc overlap)
together with the new-user friction issues found while reviewing the
section in detail.

README:

* Remove a stray AI-generated closing line ("How's that? Let me know
  if there's anything else...") that had landed in the "Why Peroxide?"
  section.
* Drop five subsections that duplicate cargo doc or are no longer
  useful: "Latest README version", "Useful tips for features",
  "Module Structure", "Documentation" (just a badge already shown at
  the top), and 
[…]
