+++ README.md
@@ -21,21 +21,15 @@ Rust numeric library contains linear algebra, numerical analysis, statistics and
     - [5. Batteries included](#5-batteries-included)
     - [6. Compatible with Mathematics](#6-compatible-with-mathematics)
     - [7. Written in Rust](#7-written-in-rust)
-  - [Latest README version](#latest-readme-version)
   - [Pre-requisite](#pre-requisite)
   - [Install](#install)
-    - [Basic Installation](#basic-installation)
-    - [Featured Installation](#featured-installation)
-    - [Available Features](#available-features)
-    - [Install Examples](#install-examples)
-  - [Useful tips for features](#useful-tips-for-features)
-  - [Module Structure](#module-structure)
-  - [Documentation](#documentation)
+    - [Most common combinations](#most-common-combinations)
+    - [Hello, Peroxide](#hello-peroxide)
+    - [Available features](#available-features)
   - [Examples](#examples)
-  - [Release Info](#release-info)
-  - [Contributing Guide](#contributing-guide)
-  - [LICENSE](#license)
-  - [TODO](#todo)
+  - [Release notes](#release-notes)
+  - [Contributing](#contributing)
+  - [License](#license)
   - [Cite Peroxide](#cite-peroxide)
 
 ## Why Peroxide?
@@ -175,7 +169,7 @@ Peroxide can do many things.
 - Automatic Differentiation
   - Const-generic `Jet<N>` type for arbitrary-order forward AD
   - Type aliases: `Dual` (1st order), `HyperDual` (2nd order)
-  - Normalized Taylor coefficients — no binomial overhead
+  - Normalized Taylor coefficients (no binomial overhead)
   - `#[ad_function]` proc macro for automatic gradient/hessian generation
   - Exact Jacobian via `jacobian()` function
   - `Real` trait to constrain for `f64` and `Jet<N>`
@@ -190,35 +184,16 @@ Peroxide can do many things.
   - Non-linear regression
     - Gradient Descent
     - Levenberg Marquardt
-  - Ordinary Differential Equation
-    - Trait based ODE solver (after `v0.36.0`)
-    - Explicit integrator
-      - Ralston's 3rd order
-      - Runge-Kutta 4th order
-      - Ralston's 4th order
-      - Runge-Kutta 5th order
-    - Embedded integrator
-      - Bogacki-Shampine 3(2)
-      - Runge-Kutta-Fehlberg 5(4)
-      - Dormand-Prince 5(4)
-      - Tsitouras 5(4)
-      - Runge-Kutta-Fehlberg 8(7)
-    - Implicit integrator
-      - Gauss-Legendre 4th order
+  - Ordinary Differential Equation (trait-based since `v0.36.0`)
+    - Explicit: Ralston 3rd / Runge-Kutta 4th / Ralston 4th / Runge-Kutta 5th
+    - Embedded: Bogacki-Shampine 3(2) / Runge-Kutta-Fehlberg 5(4) / Dormand-Prince 5(4) / Tsitouras 5(4) / Runge-Kutta-Fehlberg 8(7)
+    - Implicit: Gauss-Legendre 4th order
   - Numerical Integration
     - Newton-Cotes Quadrature
     - Gauss-Legendre Quadrature (up to 30 order)
-    - Gauss-Kronrod Quadrature (Adaptive)
-      - G7K15, G10K21, G15K31, G20K41, G25K51, G30K61
-    - Gauss-Kronrod Quadrature (Relative tolerance)
-      - G7K15R, G10K21R, G15K31R, G20K41R, G25K51R, G30K61R
-  - Root Finding
-    - Trait based root finding (after `v0.37.0`)
-    - Bisection
[…]

[…]
