### 2026-05-15 a64a557
TEST: Strengthen optimize and integral test suites (JOSS ariostas #3)
Reviewer ariostas noted that `tests/optimize.rs` only printed the
parameter vector without asserting on it ("if the algorithm changes
by accident, it wouldn't be detectable") and that `tests/integral.rs`
exercised only a single simple integral with a single quadrature
method.

tests/optimize.rs (2 -> 3 tests, 0 -> 8 assertions):

* `test_LM_quadratic_recovers_parameters` keeps the existing quadratic
  fit but now asserts that every parameter is within 1e-9 of the
  ground truth (probe shows actual error is bit-exact, so the 
[…]
