+++ tests/integral.rs
@@ -1,15 +1,161 @@
+//! Regression tests for `integrate(...)` (`tests/integral.rs`).
+//!
+//! The original version only exercised `int x.cos() dx` over [0, 1]
+//! with `G7K15R`. These tests cover several integrand families
+//! (polynomial, trigonometric, exponential, Gaussian, rational) and
+//! sweep every Gauss-Kronrod tolerance variant so that an accidental
+//! change to any one quadrature path is caught.
+
+use std::f64::consts::{E, PI};
+
 use peroxide::fuga::*;
 
-#[test]
-fn test_cosine_integral() {
-    let a = 0.0;
-    let b = 1.0;
-    let result = integrate(|x| x.cos(), (a, b), G7K15R(1e-8, 20));
-    let expected = 1f64.sin();
+/// Tolerance for smooth, low-amplitude integrands. The adaptive
+/// Gauss-Kronrod methods converge to machine epsilon on the test
+/// integrands; the tolerance leaves a comfortable safety margin for
+/// platform-specific rounding without masking a real regression.
+const SMOOTH_TOL: f64 = 1e-10;
+
+fn assert_close(name: &str, method: &str, got: f64, expected: f64, tol: f64) {
+    let err = (got - expected).abs();
     assert!(
-        (result - expected).abs() < 1e-10,
-        "Expected: {}, Got: {}",
+        err < tol,
+        "{}: {} gave {:.17e}, expected {:.17e}, |err| = {:.2e}",
+        name,
+        method,
+        got,
         expected,
-        result
+        err,
+    );
+}
+
+/// All adaptive Gauss-Kronrod variants exposed by the library.
+fn gk_methods(tol: f64) -> [(&'static str, Integral); 6] {
+    [
+        ("G7K15R", G7K15R(tol, 20)),
+        ("G10K21R", G10K21R(tol, 20)),
+        ("G15K31R", G15K31R(tol, 20)),
+        ("G20K41R", G20K41R(tol, 20)),
+        ("G25K51R", G25K51R(tol, 20)),
+        ("G30K61R", G30K61R(tol, 20)),
+    ]
+}
+
+#[test]
+fn test_polynomial() {
+    // int_0^1 x^3 dx = 1/4
+    let expected = 0.25_f64;
+    for (name, method) in gk_methods(1e-8) {
+        let got: f64 = integrate(|x: f64| x.powi(3), (0.0, 1.0), method);
+        assert_close("int x^3 over [0,1]", name, got, expected, SMOOTH_TOL);
+    }
+}
+
+#[test]
+fn test_trig() {
+    // int_0^1 cos(x) dx = sin(1)
+    let expected = 1f64.sin();
+    for (name, method) in gk_methods(1e-8) {
+        let got: f64 = integrate(|x: f64| x.cos(), (0.0, 1.0), method);
+        assert_close("int cos over [0,1]", name, got, expected, SMOOTH_TOL);
+    }
+
+    // int_0^pi sin(x) dx = 2
+    let expected = 2.0_f64;
+    for (name, method) in gk_methods(1e-8) {
+        let got: f64 = integrate(|x: f64| x.sin(), (0.0, PI), method);
+        assert_close("int sin over [0,pi]", name, got, expected, SMOOTH_TOL);
+    }
+}
+
+#[test]
+fn test_exponential() {
+    // int_0^2 exp(x) dx = exp(2) - 1
+    let expected = E * E - 1.0;
+    for (name, method) in gk_methods(1e-8) {
+        let got: f64 = integrate(|x: f64| x.exp(), (0.0, 2.0), method);
+        assert_close("int exp over [0,2]", name, got, expected, SMOOTH_TOL);
+    }
+}
+
+#[test]
+fn test_gaussian() {
[…]
