### 2026-05-15 0976ee4
TEST: Trim narration and match existing peroxide test style
The first version of these tests opened with a module-level summary,
described each assertion with several lines of motivation, and added
doc comments to the private helpers. That kind of running commentary
sits awkwardly next to the rest of the suite (`tests/jet_tests.rs`,
`tests/numerical.rs`, `tests/ad.rs`, etc.), which states facts in one
or two lines and lets the test name carry the intent.

* Drop the `//!` headers on both files.
* Replace the multi-line rationale comments inside each test with at
  most one short formula-style 
[…]
