+++ tests/integral.rs
@@ -1,161 +1,104 @@
-//! Regression tests for `integrate(...)` (`tests/integral.rs`).
-//!
-//! The original version only exercised `int x.cos() dx` over [0, 1]
-//! with `G7K15R`. These tests cover several integrand families
-//! (polynomial, trigonometric, exponential, Gaussian, rational) and
-//! sweep every Gauss-Kronrod tolerance variant so that an accidental
-//! change to any one quadrature path is caught.
-
-use std::f64::consts::{E, PI};
+use std::f64::consts::{E, FRAC_PI_2, PI};
 
 use peroxide::fuga::*;
 
-/// Tolerance for smooth, low-amplitude integrands. The adaptive
-/// Gauss-Kronrod methods converge to machine epsilon on the test
-/// integrands; the tolerance leaves a comfortable safety margin for
-/// platform-specific rounding without masking a real regression.
-const SMOOTH_TOL: f64 = 1e-10;
+const TOL: f64 = 1e-10;
 
-fn assert_close(name: &str, method: &str, got: f64, expected: f64, tol: f64) {
-    let err = (got - expected).abs();
+fn assert_close(actual: f64, expected: f64) {
     assert!(
-        err < tol,
-        "{}: {} gave {:.17e}, expected {:.17e}, |err| = {:.2e}",
-        name,
-        method,
-        got,
-        expected,
-        err,
+        (actual - expected).abs() < TOL,
+        "actual = {:.17e}, expected = {:.17e}, diff = {:.3e}",
+        actual, expected, (actual - expected).abs()
     );
 }
 
-/// All adaptive Gauss-Kronrod variants exposed by the library.
-fn gk_methods(tol: f64) -> [(&'static str, Integral); 6] {
+fn gk_methods(tol: f64) -> [Integral; 6] {
     [
-        ("G7K15R", G7K15R(tol, 20)),
-        ("G10K21R", G10K21R(tol, 20)),
-        ("G15K31R", G15K31R(tol, 20)),
-        ("G20K41R", G20K41R(tol, 20)),
-        ("G25K51R", G25K51R(tol, 20)),
-        ("G30K61R", G30K61R(tol, 20)),
+        G7K15R(tol, 20),
+        G10K21R(tol, 20),
+        G15K31R(tol, 20),
+        G20K41R(tol, 20),
+        G25K51R(tol, 20),
+        G30K61R(tol, 20),
     ]
 }
 
 #[test]
 fn test_polynomial() {
     // int_0^1 x^3 dx = 1/4
-    let expected = 0.25_f64;
-    for (name, method) in gk_methods(1e-8) {
-        let got: f64 = integrate(|x: f64| x.powi(3), (0.0, 1.0), method);
-        assert_close("int x^3 over [0,1]", name, got, expected, SMOOTH_TOL);
+    for m in gk_methods(1e-8) {
+        let r: f64 = integrate(|x: f64| x.powi(3), (0.0, 1.0), m);
+        assert_close(r, 0.25);
     }
 }
 
 #[test]
 fn test_trig() {
-    // int_0^1 cos(x) dx = sin(1)
-    let expected = 1f64.sin();
-    for (name, method) in gk_methods(1e-8) {
-        let got: f64 = integrate(|x: f64| x.cos(), (0.0, 1.0), method);
-        assert_close("int cos over [0,1]", name, got, expected, SMOOTH_TOL);
+    // int_0^1 cos x dx = sin(1)
+    for m in gk_methods(1e-8) {
+        let r: f64 = integrate(|x: f64| x.cos(), (0.0, 1.0), m);
+        assert_close(r, 1f64.sin());
     }
-
-    // int_0^pi sin(x) dx = 2
-    let expected = 2.0_f64;
[…]
