+++ tests/optimize.rs
@@ -1,39 +1,18 @@
-//! Regression tests for `Optimizer` (`tests/optimize.rs`).
-//!
-//! The original version only called `.optimize()` and printed the
-//! returned parameter vector, which would not have caught a silent
-//! regression in the optimizer itself. These tests assert the
-//! parameters that the fit should recover.
-
 extern crate peroxide;
 use peroxide::{fuga::*, hstack};
 
 const P_TRUE: [f64; 3] = [1.0, 2.0, 3.0];
 
-// Quadratic model y = p[0] * t^2 + p[1] * t + p[2].
-//
-// Signature must stay `&Vec<f64>` to satisfy `Optimizer`'s trait bound.
+// y = p[0] * t^2 + p[1] * t + p[2]
 #[allow(clippy::ptr_arg)]
 fn quadratic(x: &Vec<f64>, p: Vec<AD>) -> Option<Vec<AD>> {
-    Some(
-        x.iter()
-            .map(|t| AD1(*t, 0f64))
-            .map(|t| p[0] * t.powi(2) + p[1] * t + p[2])
-            .collect(),
-    )
+    Some(x.iter().map(|t| AD1(*t, 0f64)).map(|t| p[0] * t.powi(2) + p[1] * t + p[2]).collect())
 }
 
-// Exponential model y = p[0] * exp(p[1] * t).
-//
-// Signature must stay `&Vec<f64>` to satisfy `Optimizer`'s trait bound.
+// y = p[0] * exp(p[1] * t)
 #[allow(clippy::ptr_arg)]
 fn exp_decay(x: &Vec<f64>, p: Vec<AD>) -> Option<Vec<AD>> {
-    Some(
-        x.iter()
-            .map(|t| AD1(*t, 0f64))
-            .map(|t| p[0] * (p[1] * t).exp())
-            .collect(),
-    )
+    Some(x.iter().map(|t| AD1(*t, 0f64)).map(|t| p[0] * (p[1] * t).exp()).collect())
 }
 
 #[test]
@@ -44,27 +23,15 @@ fn test_LM_quadratic_recovers_parameters() {
     let data = hstack!(x, y);
 
     let mut opt = Optimizer::new(data, quadratic);
-    let p_est = opt
-        .set_init_param(vec![1f64, 1f64, 1f64])
+    let p = opt.set_init_param(vec![1f64, 1f64, 1f64])
         .set_max_iter(50)
         .set_method(LevenbergMarquardt)
         .set_lambda_init(1e-3)
         .set_lambda_max(1e+3)
         .optimize();
 
-    // Levenberg-Marquardt on a linear-in-parameters least-squares
-    // problem should converge to the exact solution within a handful
-    // of iterations. A tight tolerance is appropriate here.
-    for (i, expected) in P_TRUE.iter().enumerate() {
-        let err = (p_est[i] - expected).abs();
-        assert!(
-            err < 1e-9,
-            "LM did not recover parameter {i}: got {got}, expected {expected}, |err| = {err:.2e}",
-            i = i,
-            got = p_est[i],
-            expected = expected,
-            err = err,
-        );
+    for i in 0..3 {
+        assert!((p[i] - P_TRUE[i]).abs() < 1e-9);
     }
 }
 
@@ -74,77 +41,40 @@ fn test_GD_quadratic_makes_progress() {
     let x = seq(0, 10, 0.1);
     let y = x.fmap(|t| P_TRUE[0] * t.powi(2) + P_TRUE[1] * t + P_TRUE[2]);
     let data = hstack!(x, y);
-
     let p_init = [1f64, 1f64, 1f64];
+
     let mut opt = Optimizer::new(data, quadratic);
-    let p_est = opt
[…]
