### 2026-05-15 bb1d636
PAPER: Add citations and trade-off discussion (JOSS jonaspleyer #100)
Address all four checkboxes in jonaspleyer's paper review:

* Cite LAPACK [@lapack] (Anderson et al., LAPACK Users' Guide, 3rd
  ed., SIAM 1999) at the single body mention of LAPACK on line 61.
* Cite ndarray [@ndarray] (rust-ndarray developers, v0.17.2) in the
  Table 1 caption and in the new architecture trade-off sentence.
* Add a one-sentence trade-off note to the Architecture paragraph
  spelling out the two known downsides of the flat Vec<f64> layout
  (compile-time dimension checking versus nalgebra; N-dimensional
  g
[…]
