+++ paper/paper.bib
@@ -303,3 +303,40 @@
   url     = {https://parquet.apache.org/},
   year    = {2026}
 }
+
+@book{lapack,
+  author    = {Anderson, E. and Bai, Z. and Bischof, C. and Blackford, S. and Demmel, J. and Dongarra, J. and Du Croz, J. and Greenbaum, A. and Hammarling, S. and McKenney, A. and Sorensen, D.},
+  title     = {{LAPACK} Users' Guide},
+  edition   = {Third},
+  publisher = {Society for Industrial and Applied Mathematics},
+  address   = {Philadelphia, PA},
+  year      = {1999},
+  isbn      = {0-89871-447-8},
+  doi       = {10.1137/1.9780898719604}
+}
+
+@software{openblas,
+  author  = {{OpenBLAS developers}},
+  title   = {{OpenBLAS}: An optimized {BLAS} library},
+  url     = {https://github.com/OpenMathLib/OpenBLAS},
+  version = {0.3.33},
+  year    = {2026}
+}
+
+@software{ndarray,
+  author  = {{rust-ndarray developers}},
+  title   = {{ndarray}: an n-dimensional array for general elements and for numerics in {Rust}},
+  url     = {https://github.com/rust-ndarray/ndarray},
+  version = {0.17.2},
+  year    = {2026}
+}
+
+@inproceedings{enzyme,
+  author    = {Moses, William S. and Churavy, Valentin},
+  title     = {Instead of rewriting foreign code for machine learning, automatically synthesize fast gradients},
+  booktitle = {Advances in Neural Information Processing Systems},
+  volume    = {33},
+  year      = {2020},
+  pages     = {12472--12485},
+  url       = {https://arxiv.org/abs/2010.01709}
+}
