+++ paper/paper.md
@@ -45,9 +45,9 @@ bibliography: paper.bib
 
 # Summary
 
-Peroxide is a numerical computing library for the Rust programming language that integrates eight domains within a single crate: linear algebra, ordinary differential equation (ODE) solving, adaptive quadrature, spline interpolation with exact polynomial calculus, root-finding, forward-mode automatic differentiation (AD), statistics, and multi-format DataFrame I/O.
+Peroxide is a Rust numerical computing library that integrates eight domains in one crate: linear algebra, ordinary differential equation (ODE) solving, adaptive quadrature, spline interpolation with exact polynomial calculus, root-finding, forward-mode automatic differentiation (AD), statistics, and multi-format DataFrame I/O.
 We designed Peroxide for applied scientists, engineers, and students who need a cohesive numerical environment in Rust without assembling a multi-crate dependency graph.
-A dual-module API provides progressive disclosure: the `prelude` module offers sensible defaults, while the `fuga` module exposes explicit algorithm selection, addressing Rust's lack of default function arguments.
+A dual-module API offers `prelude` defaults and a `fuga` module for explicit algorithm selection, addressing Rust's lack of default function arguments.
 Peroxide is released under the MIT/Apache-2.0 dual license and archived on Zenodo [@peroxide].
 
 # Statement of need
@@ -58,12 +58,12 @@ In Python, SciPy [@scipy] and NumPy [@numpy] consolidate these capabilities unde
 In Rust, they are scattered across specialized crates, each with its own type conventions and trait hierarchies.
 
 \autoref{tab:comparison} surveys the Rust numerical ecosystem.
-Specialist crates excel individually: diffsol [@diffsol] provides ODE/DAE solvers with sensitivity analysis; faer [@faer] offers pure-Rust linear algebra competitive with LAPACK; nalgebra [@nalgebra] provides matrix abstractions widely adopted across the ecosystem; argmin [@argmin] supplies over 20 optimization algorithms; Polars [@polars] dominates tabular data processing.
+Specialist crates excel individually: diffsol [@diffsol] provides ODE/DAE solvers with sensitivity analysis; faer [@faer] offers pure-Rust linear algebra competitive with LAPACK [@lapack]; nalgebra [@nalgebra] provides matrix abstractions widely adopted across the ecosystem; argmin [@argmin] supplies over 20 optimization algorithms; Polars [@polars] dominates tabular data processing.
 russell [@russell] covers the broadest range among alternatives (ODE solvers, quadrature, root-finding, and statistics) but lacks AD, spline calculus, and DataFrame support.
 No single crate spans all eight domains shown in \autoref{tab:comparison}.
 Peroxide fills this integration gap.
 
-: Feature coverage of Rust numerical libraries. Each entry reflects the crate's own API, including optional feature flags; a dash indicates the feature is absent. Crates providing only N-dimensional array abstractions without numerical algorithms (e.g., ndarray) are omitted. \label{tab:comparison}
+: Feature coverage of Rust numerical libraries. Each entry reflects the crate's own API, including optional feature flags; a dash indicates the feature is absent. Crates providing only N-dimensional array abstractions without numerical algorithms (e.g., ndarray [@ndarray]) are omitted. \label{tab:comparison}
 
 |                     | Peroxide | diffsol | russell | faer | nalgebra | argmin | num-dual | Polars |
 |---------------------|:--------:|:-------:|:-------:|:----:|:--------:|:------:|:--------:|:------:|
@@ -77,13 +77,15 @@ Peroxide fills this integration gap.
 | DataFrame I/O       | $\checkmark$        | —       | —       | —    | —        | —      | —        | $\checkmark$      |
 
 We do not claim superiority over any specialist crate within its domain.
[…]
