### 2026-05-15 1b9838c
FIX: Silence clippy::eq_op on the AD commutativity check (CI hotfix)
Clippy 1.95 on the CI runner promotes `clippy::eq_op` from warn to
deny for the commutativity assertion in examples/ad.rs:

    assert_eq!(2f64 * b, b * 2f64);

The two sides are intentionally identical — the test exists to
confirm that `2 * b == b * 2` holds for `AD` types — but the lint
normalizes both sides and reports them as a no-op comparison.

Wrap the assertion in a block-level `#[allow(clippy::eq_op)]` with
a short note. The check itself is preserved verbatim. Local
clippy run (Rust 1.93 stable) now finishes with zer
[…]
