### 2026-05-16 67ca352
RLSE: Ver 0.41.1
JOSS review cycle for openjournals/joss-reviews#10366.

- Bug fix: SPMatrix::new row_ics nnz length
- Build: O3-* convenience features for BLAS/LAPACK backend
- Docs: README restructure, Statement of need, docs.rs build,
  feature documentation, expanded CONTRIBUTING
- CI / Lint: separated feature jobs, clippy --all-targets
  blocking, clippy_verify determinism oracle
- Test: stronger optimize / integral suites
- Paper: JOSS paper sources, co-authors, citations

Refs Axect/Peroxide#97, #98, #99, #100
Addresses ariostas review items 1-5 on
openjournals/joss-reviews#10366

See R
[…]
