### 2026-05-26 Axect/Peroxide#97 @Axect
Resolved in 70d3533, shipped in v0.41.1 and v0.41.2.

Thanks @jonaspleyer for the catch. The root cause turned out to be on our side rather than a nightly-only feature gap: we had `#![cfg_attr(docsrs, feature(doc_auto_cfg))]` in `src/lib.rs`, but the `doc_auto_cfg` nightly feature was merged into the always-on `doc_cfg` machinery in Rust 1.92 (RFC 3631), so the gate itself stopped compiling. Removing the `cfg_attr` (and the now-unused `--cfg docsrs` rustdoc arg) unblocked docs.rs entirely.

The v0.41.2 page is live at <https://docs.rs/peroxide/0.41.2>.
