### 2026-05-26 Axect/Peroxide#98 @Axect
Thanks @jonaspleyer for the very thorough feature-matrix test. It exposed several latent issues that we have now addressed across the following commits, all shipped in v0.41.1:

| Symptom in your matrix          | Fix                                                          |
| ------------------------------- | ------------------------------------------------------------ |
| `O3` linker errors              | 95c42c5 : new `O3-openblas` / `O3-netlib` / `O3-accelerate` / `O3-mkl` convenience features that pull in the matching `*-src` backend so symbols resolve. |
| `tests/o3/lapack.rs` import err | 9141070 : re-export `UPLO` from `fuga`, fix `tests/blas_test` macro import. |
| `nc` / `netcdf` HDF5 failure    | d4e3c9c : documented HDF5 1.x constraint (pinned `netcdf 0.7` does not yet support HDF5 2.x); a `netcdf 0.12` migration is tracked separately. |
| `plot` doctests failing         | cd73e1b : marked the four `.savefig(...)` doctests as `no_run` since they require matplotlib + a writable target path. |
| Sparse feature documentation    | fc1c41a + d4e3c9c : rewrote the Pre-requisite / Available Features sections of the README with full install instructions per feature. |
| No CI coverage for these        | 093f4b1 : workflow now runs default / `O3-openblas` / pure-Rust / `plot` jobs in parallel, plus a blocking `clippy --all-targets`. |

The five CI jobs are all green on master at v0.41.2.
