### 2026-05-26 Axect/Peroxide#99 @Axect
Thanks @jonaspleyer, applied in 86cc03d (v0.41.1), with follow-ups in 8901470 (`documentation` link to docs.rs) and 63c881c (crate landing page Statement of need).

Following your suggestions, the README now:

- Drops the cargo-doc-overlapping subsections (`Latest README version`,  `Useful tips for features`, `Module Structure`, `Documentation`,  `Examples`, `Release Info`, `Contributing Guide`, `LICENSE`, `TODO`), with the surviving pointers consolidated into a short footer that links out to docs.rs / `examples/` / `CONTRIBUTING.md` / `LICENSE`.
- Adds a new **Quickstart** section so a first-time reader can build and run something without scrolling.
- Adds a **Source layout** section pulled from `CONTRIBUTING.md` so contributors can see the module structure without leaving the README. Keeps the per-component feature lists nested (they double as the "what's in the box" inventory for users picking the crate), but each section is now shorter and groups variants where natural.

Diff: <https://github.com/Axect/Peroxide/commit/86cc03d>. Happy to trim further if any of the remaining sections still feel redundant with docs.rs.
