### 2026-05-26 Axect/Peroxide#100 @Axect
Thanks @jonaspleyer. All four checkboxes addressed in bb1d636 (v0.41.1):

- [x] **LAPACK citation**: added `@lapack` (Anderson et al., *LAPACK Users' Guide*, 3rd ed., SIAM 1999) at the single body mention.
- [x] **`ndarray` citation**: added `@ndarray` to the Table 1 caption and to the new trade-off discussion.
- [x] **Matrix memory layout trade-off**: added a paragraph in  section 3 discussing the `Vec<f64>` row-major layout vs. the strided-view approach used by `ndarray` / `nalgebra`: contiguous storage gives us BLAS-friendly cache locality and a simple FFI boundary, at the cost of cheap non-contiguous views (transposed slices, fancy indexing) which the strided crates handle natively.
- [x] **Enzyme comparison**: added to section 4 (AD section). The contrast is intentional: Enzyme operates at the LLVM-IR level and can differentiate arbitrary Rust (and other LLVM-emitting languages), giving it broader reach but a heavier compile-time pipeline, while Peroxide's const-generic `Jet<N>` is a source-level forward-mode primitive with zero external tooling, a smaller scope picked for ergonomics and reverse-incremental compile times.
