### 2026-06-09 ec37e61
STYLE: Apply cargo fmt to files added after PR #96 (JOSS ariostas)
