### 2026-06-09 bd36784
CI: Add rustfmt --check job (JOSS ariostas)
