### 2026-06-15 Axect/Peroxide#99 @jonaspleyer
I like the addition of the `Quickstart` section. That makes a lot of sense to get a first glimpse of what the code is supposed to look and "feel" like. I don't really see how the `Source Layout` section improves the README. My opinion is that if anyone wants to deal with the code, they will clone the repository and look inside the folders anyways. Otherwise only leave a high-level description in the README. Furthermore, every time that a new file is added/removed this structure needs to be updated as well. Therefore I find it more of a burden rather than an improvement.

Addendum: Of course I also like that you dropped the duplicated stuff.
