### 2026-06-15 Axect/Peroxide#98 @jonaspleyer
One clear downside of having many features is that in principle each individual feature and also all combinations should be tested within a CI. what is your strategy to prevent this from creeping up again in the future?
