### 2026-06-16 Axect/Peroxide#98 @jonaspleyer
I ran the tests from above again. And this time I actually got more errors.
I am running Archlinux and have followed https://github.com/Axect/Peroxide_BLAS to install blas.

  <summary>
    ArchLinux Info & Rust Version
  </summary>

```text
# uname -a
Linux RETRACTED 6.18.28-1-lts #1 SMP PREEMPT_DYNAMIC Fri, 08 May 2026 07:04:32 +0000 x86_64 GNU/Linux
```

```text
# pacman -Qs blas
local/blas 3.12.1-2
    Basic Linear Algebra Subprograms
local/blas64 3.12.1-2
    Basic Linear Algebra Subprograms (64-bit integers)
local/cblas 3.12.1-2
    C interface to BLAS
local/cblas64 3.12.1-2
    C interface to BLAS (64-bit integers)
local/cuda 13.2.1-2
    NVIDIA's GPU programming toolkit
local/openblas 0.3.33-1
    An optimized BLAS library based on GotoBLAS2 1.13 BSD
```

```text
# pacman -Qs lapack
local/lapack 3.12.1-2
    Linear Algebra PACKage
```

```text
# rustup show
Default host: x86_64-unknown-linux-gnu
rustup home:  /home/jonas/.rustup

installed toolchains
--------------------
stable-x86_64-unknown-linux-gnu (active, default)
nightly-x86_64-unknown-linux-gnu
nightly-2025-02-01-x86_64-unknown-linux-gnu
nightly-2025-06-01-x86_64-unknown-linux-gnu
nightly-2026-04-03-x86_64-unknown-linux-gnu
1.92-x86_64-unknown-linux-gnu

active toolchain
----------------
name: stable-x86_64-unknown-linux-gnu
active because: it's the default toolchain
installed targets:
  aarch64-unknown-linux-gnu
  thumbv6m-none-eabi
  wasm32-unknown-unknown
  x86_64-apple-darwin
  x86_64-unknown-linux-gnu
```

The results below show that many more combinations seem to not be working as intended. It took 1h 8min on my machine (Ryzen 5950X, 16 Core).

  <summary>
    Test results
  </summary>

```
✅ = Passed
❌ = Failed
⏩ = Skipped (due to previous Fail)
1st col: debug mode
2nd col: release mode
```

```text
# Testing Examples
✅✅ ad
✅✅ dataframe
✅✅ gradient
✅✅ lorenz_gl4
✅✅ norm
✅✅ ode_test_rkf78
✅✅ root_test
✅✅ area_under_spline
✅✅ det
✅✅ hessian
✅✅ lorenz_rk4
✅✅ ode_env
✅✅ optim
✅✅ solve
✅✅ broyden_test
✅✅ df_csv
✅✅ higher_order_ad
✅✅ lorenz_rkf45
✅✅ ode_test_gl4
✅✅ poly
✅✅ spread
✅✅ bspline_calculus
✅✅ df_nc
✅✅ integrate
✅✅ lorenz_rkf78
✅✅ ode_test_orbit
✅✅ pseudo_inv
✅✅ svd
✅✅ b_spline_test
✅✅ df_parquet
✅✅ jet_ad
✅✅ lorenz_tsit45
✅✅ ode_test_rk4
✅✅ real_trait_test
✅✅ thomas
✅✅ chebyshev
✅✅ dist
✅✅ lorenz_dp45
✅✅ matmul
✅✅ ode_test_rkf45
✅✅ root_macro_test
# Feature Combinations: 1
✅✅ default
❌❌ O3
✅✅ arrow
✅✅ blas
✅✅ complex
✅✅ csv
✅✅ indexmap
✅✅ json
✅✅ lapack
❌❌ nc
❌❌ netcdf
✅✅ num-complex
✅✅ parallel
✅✅ parquet
✅✅ plot
✅✅ pyo3
✅✅ rayon
✅✅ rkyv
✅✅ serde
# Feature Combinations: 2
⏩   default,O3
✅✅ default,arrow
✅✅ default,blas
✅✅ default,complex
✅✅ default,csv
✅✅ default,indexmap
✅✅ default,json
✅✅ default,lapack
⏩   default,nc
⏩   default,netcdf
✅✅ default,num-complex
✅✅ default,parallel
✅✅ default,parquet
✅✅ default,plot
✅✅ default,pyo3
✅✅ default,rayon
✅✅ default,rkyv
✅✅ default,serde
⏩   O3,arrow
⏩   O3,blas
⏩   O3,complex
⏩   O3,csv
⏩   O3,indexmap
⏩   O3,json
⏩   O3,lapack
⏩   O3,nc
⏩   O3,netcdf
⏩   O3,num-complex
⏩   O3,parallel
⏩   O3,parquet
⏩   O3,plot
⏩   O3,pyo3
⏩   O3,rayon
⏩   O3,rkyv
⏩   O3,serde
✅✅ arrow,blas
✅✅ arrow,complex
✅✅ arrow,csv
✅✅ arrow,indexmap
✅✅ arrow,json
✅✅ arrow,lapack
⏩   arrow,nc
⏩   arrow,netcdf
✅✅ arrow,num-complex
✅✅ arrow,parallel
✅✅ arrow,parquet
✅✅ arrow,plot
✅✅ arrow,pyo3
✅✅ arrow,rayon
✅✅ arrow,rkyv
✅✅ arrow,serde
✅✅ blas,complex
❌❌ blas,csv
❌❌ blas,indexmap
❌❌ blas,json
❌❌ blas,lapack
⏩   blas,nc
⏩   blas,netcdf
❌❌ blas,num-complex
❌❌ blas,parallel
❌❌ blas,parquet
❌❌ blas,plot
❌❌ blas,pyo3
❌❌ blas,rayon
❌❌ blas,rkyv
❌❌ blas,serde
❌❌ complex,csv
❌❌ complex,indexmap
❌❌ complex,json
❌❌ complex,lapack
⏩   complex,nc
⏩   complex,netcdf
✅✅ complex,num-complex
❌❌ complex,parallel
❌❌ complex,parquet
❌❌ complex,plot
❌❌ complex,pyo3
❌❌ complex,rayon
❌❌ complex,rkyv
❌❌ complex,serde
❌❌ csv,indexmap
❌❌ csv,json
❌❌ csv,lapack
⏩   csv,nc
⏩   csv,netcdf
❌❌ csv,num-complex
❌❌ csv,parallel
❌❌ csv,parquet
❌❌ csv,plot
❌❌ csv,pyo3
❌❌ csv,rayon
❌❌ csv,rkyv
❌❌ csv,serde
❌❌ indexmap,json
❌❌ indexmap,lapack
⏩   indexmap,nc
⏩   indexmap,netcdf
❌❌ indexmap,num-complex
❌❌ indexmap,parallel
❌❌ indexmap,parquet
❌❌ indexmap,plot
❌❌ indexmap,pyo3
❌❌ indexmap,rayon
❌❌ indexmap,rkyv
❌❌ indexmap,serde
❌❌ json,lapack
⏩   json,nc
⏩   json,netcdf
❌❌ json,num-complex
❌❌ json,parallel
❌❌ json,parquet
❌❌ json,plot
❌❌ json,pyo3
❌❌ json,rayon
❌❌ json,rkyv
❌❌ json,serde
⏩   lapack,nc
⏩   lapack,netcdf
❌❌ lapack,num-complex
❌❌ lapack,parallel
❌❌ lapack,parquet
❌❌ lapack,plot
❌❌ lapack,pyo3
❌❌ lapack,rayon
❌❌ lapack,rkyv
❌❌ lapack,serde
⏩   nc,netcdf
⏩   nc,num-complex
⏩   nc,parallel
⏩   nc,parquet
⏩   nc,plot
⏩   nc,pyo3
⏩   nc,rayon
⏩   nc,rkyv
⏩   nc,serde
⏩   netcdf,num-complex
⏩   netcdf,parallel
⏩   netcdf,parquet
⏩   netcdf,plot
⏩   netcdf,pyo3
⏩   netcdf,rayon
⏩   netcdf,rkyv
⏩   netcdf,serde
❌❌ num-complex,parallel
❌❌ num-complex,parquet
❌❌ num-complex,plot
❌❌ num-complex,pyo3
❌❌ num-complex,rayon
❌❌ num-complex,rkyv
❌❌ num-complex,serde
❌❌ parallel,parquet
❌❌ parallel,plot
❌❌ parallel,pyo3
✅✅ parallel,rayon
❌❌ parallel,rkyv
❌❌ parallel,serde
❌❌ parquet,plot
❌❌ parquet,pyo3
❌❌ parquet,rayon
❌❌ parquet,rkyv
❌❌ parquet,serde
✅✅ plot,pyo3
❌❌ plot,rayon
❌❌ plot,rkyv
❌❌ plot,serde
❌❌ pyo3,rayon
❌❌ pyo3,rkyv
❌❌ pyo3,serde
❌❌ rayon,rkyv
❌❌ rayon,serde
❌❌ rkyv,serde
```

Here is the script that I used to run the tests:

  <summary>
    Python Script
  </summary>

```python
import itertools
import subprocess

examples = [
    "ad",
    "dataframe",
    "gradient",
    "lorenz_gl4",
    "norm",
    "ode_test_rkf78",
    "root_test",
    "area_under_spline",
    "det",
    "hessian",
    "lorenz_rk4",
    "ode_env",
    "optim",
    "solve",
    "broyden_test",
    "df_csv",
    "higher_order_ad",
    "lorenz_rkf45",
    "ode_test_gl4",
    "poly",
    "spread",
    "bspline_calculus",
    "df_nc",
    "integrate",
    "lorenz_rkf78",
    "ode_test_orbit",
    "pseudo_inv",
    "svd",
    "b_spline_test",
    "df_parquet",
    "jet_ad",
    "lorenz_tsit45",
    "ode_test_rk4",
    "real_trait_test",
    "thomas",
    "chebyshev",
    "dist",
    "lorenz_dp45",
    "matmul",
    "ode_test_rkf45",
    "root_macro_test",
]

features = [
    "default",
    "O3",
    "arrow",
    "blas",
    "complex",
    "csv",
    "indexmap",
    "json",
    "lapack",
    "nc",
    "netcdf",
    "num-complex",
    "parallel",
    "parquet",
    "plot",
    "pyo3",
    "rayon",
    "rkyv",
    "serde",
]

def s(ret):
    return "✅" if ret.returncode == 0 else "❌"

# EXAMPLES
print("# Testing Examples")
for ex in examples:
    ret1 = subprocess.run(
        f"cargo run --example {ex} > /dev/null 2> /dev/null", shell=True
    )
    ret2 = subprocess.run(
        f"cargo run --release --example {ex} > /dev/null 2> /dev/null", shell=True
    )

    print(f"{s(ret1)}{s(ret2)} {ex}")

# FEATURES

def test_feature(feature_comb):
    ret1 = subprocess.run(
        f"cargo test\
            --features {feature_comb}\
            -- --test-threads=4\
            > /dev/null 2> /dev/null",
        shell=True,
    )
    ret2 = subprocess.run(
        f"cargo test\
            --release\
            --features {feature_comb}\
            -- --test-threads=4\
            > /dev/null 2> /dev/null",
        shell=True,
    )
    return (ret1, ret2)

def check_errors(comb, errors):
    for e in errors:
        if all(ei in comb for ei in e):
            return True

    return False

errors = []
for i in range(1, 4):
    print(f"# Feature Combinations: {i}")
    for comb in itertools.combinations(features, i):
        feature_comb = ",".join(comb)

        if check_errors(comb, errors):
            print(f"⏩   {feature_comb}")
            continue

        (ret1, ret2) = test_feature(feature_comb)

        if ret1.returncode != 0 or ret1.returncode != 0:
            errors.append(comb)
        print(f"{s(ret1)}{s(ret2)} {feature_comb}")
```
