### 2026-06-17 fcfd012
CI: Add cargo-hack feature-combinations job (JOSS #98)
Add a feature-combinations job that runs cargo-hack:
- build --each-feature --optional-deps: every feature builds on its own
- build --feature-powerset --depth 2 --optional-deps: pairwise coverage

Heavy compile units (arrow/parquet) and external-toolchain features
(O3 backends, nc/netcdf, plot/pyo3) are excluded from the combinatorial
job so the runner's target dir stays bounded (peak ~13 GB locally).
