### 2026-06-17 Axect/Peroxide#98 @Axect
Thanks for re-running the matrix and for the detailed environment dump.

**On the CI strategy question.** I added a dedicated `feature-combinations` job to the workflow using [`cargo-hack`](https://github.com/taiki-e/cargo-hack):

- `cargo hack build --each-feature --optional-deps` builds every feature (including the optional-dependency features) on its own.
- `cargo hack build --feature-powerset --depth 2 --optional-deps` covers every pair of the remaining features.

To keep this job a code check rather than an environment check, a few features are excluded from it: the HDF5 features (`nc` / `netcdf`) need HDF5 headers, the non-OpenBLAS backends (`O3-accelerate` / `O3-mkl` / `O3-netlib`) need a vendor toolchain or a non-Linux platform, and `arrow` / `parquet` are dropped from the powerset only because they are large compile units. The `O3-openblas` and `plot` / `pyo3` paths keep their existing dedicated jobs with the proper system setup. Locally the new job runs 94 feature configurations with a peak `target/` of about 13 GB, so it stays within a runner's disk budget. This should catch feature-gating regressions automatically from now on.

**On the new failures in your second run.** I could not reproduce any of them. I reproduced your matrix on Arch Linux (I'm also arch user 😉 ; same feature combinations, debug + release builds): the only failures were `O3`, `nc` and `netcdf` (the known backend / HDF5 cases), and every combination you marked as failing in the second run, including `blas,csv`, passed here (136 pass / 3 fail / 51 skipped).

Two things worth noting:

1. The bare `blas` / `lapack` features only enable the FFI binding crates. Peroxide's actual BLAS/LAPACK calls are all gated behind `O3`, and `blas-sys` does not link any system library on its own, so a combination like `blas,csv` does not link or call BLAS at all and cannot fail for a BLAS-related reason.
2. The failure pattern in your second run looks like resource exhaustion rather than a code regression: the failures start partway through, and from then on only the combinations that need no fresh compilation survive (`complex,num-complex`, `parallel,rayon`, `plot,pyo3`, which are alias supersets of a single feature). Running the full matrix is very disk-heavy: one pass here grew the `target/` directory by about 270 GB (to over 300 GB). If the partition filled up mid-run, every fresh build afterwards would fail exactly like this.

Could you re-run with the output captured to a log instead of `/dev/null`, and check `df -h` / `du -sh target` during the run? That would let us see the actual error and confirm whether it is disk space. The script currently discards stderr, so the real cause is not visible in the table.
