### 2026-06-24 1b6170f
feat: add shape method to MatrixTrait and implementations for ComplexMatrix and Matrix
