### 2026-07-01 Axect/Peroxide#98 @jonaspleyer
Unfortunately the `cargo hack` command does not run for me. It fails with the error message ```error: library kind `framework` is only supported on Apple targets```. I tried some stuff but could not quickly find out how to make `cargo hack` work on arch.

> On the new failures in your second run. I could not reproduce any of them. I reproduced your matrix on Arch Linux (I'm also arch user 😉 ; same feature combinations, debug + release builds): the only failures were O3, nc and netcdf (the known backend / HDF5 cases), and every combination you marked as failing in the second run, including blas,csv, passed here (136 pass / 3 fail / 51 skipped).

Maybe I misundertstood something here: So you are on Arch as well and the crate is also not building for you, although you have installed the necessary libraries that should be linked in? I am currently re-running the script as you suggested. I will report my findings.
