### 2026-07-06 72a9e56
DOCS: Promote Quickstart, condense README and CONTRIBUTING lists (JOSS #99)
Address jonaspleyer's follow-up on the review issue #99:

* Promote the "Hello, Peroxide" example to a top-level Quickstart
  section right below the badges, so the first thing a new reader
  sees is runnable code.
* Condense the "Batteries included" inventory to one line per
  category with variants grouped in parentheses (89 lines -> 15),
  as suggested in the original issue.
* Trim the CONTRIBUTING.md "Source layout" per-file tree to a
  directory-level table (one line per module), removing the
  maintenance burden 
[…]
