+++ README.md
@@ -9,6 +9,34 @@
 
 Rust numeric library contains linear algebra, numerical analysis, statistics and machine learning tools with R, MATLAB, Python like macros.
 
+## Quickstart
+
+```bash
+cargo add peroxide   # default profile is pure Rust, no system libraries needed
+```
+
+```rust
+#[macro_use]
+extern crate peroxide;
+use peroxide::fuga::*;
+
+fn main() {
+    // R / MATLAB-style matrix literals
+    let a = ml_matrix("1 2; 3 4");
+    let b = c!(5, 6);
+
+    // matrix-vector product (BLAS-dispatched when an `O3-*` feature is on)
+    let c = &a * &b;
+
+    a.print(); // pretty-formatted matrix
+    c.print(); // [17, 39]
+    a.det().print(); // -2
+    a.inv().print();
+}
+```
+
+For accelerated linear algebra, plotting, or DataFrame I/O, enable the matching feature flag (see [Install](#install) and [Available features](#available-features)).
+
 ## Table of Contents
 
 - [Peroxide](#peroxide)
@@ -24,7 +52,6 @@ Rust numeric library contains linear algebra, numerical analysis, statistics and
   - [Pre-requisite](#pre-requisite)
   - [Install](#install)
     - [Most common combinations](#most-common-combinations)
-    - [Hello, Peroxide](#hello-peroxide)
     - [Available features](#available-features)
   - [Examples](#examples)
   - [Release notes](#release-notes)
@@ -150,91 +177,20 @@ fn main() {
 
 Peroxide can do many things. 
 
-- Linear Algebra
-  - Effective Matrix structure
-  - Transpose, Determinant, Diagonal
-  - LU Decomposition, Inverse matrix, Block partitioning
-  - QR Decomposition (`O3` feature)
-  - Singular Value Decomposition (SVD) (`O3` feature)
-  - Cholesky Decomposition (`O3` feature)
-  - Reduced Row Echelon form
-  - Column, Row operations
-  - Eigenvalue, Eigenvector
-- Functional Programming
-  - Easier functional programming with `Vec<f64>`
-  - For matrix, there are three maps
-    - `fmap` : map for all elements
-    - `col_map` : map for column vectors
-    - `row_map` : map for row vectors
-- Automatic Differentiation
-  - Const-generic `Jet<N>` type for arbitrary-order forward AD
-  - Type aliases: `Dual` (1st order), `HyperDual` (2nd order)
-  - Normalized Taylor coefficients (no binomial overhead)
-  - `#[ad_function]` proc macro for automatic gradient/hessian generation
-  - Exact Jacobian via `jacobian()` function
-  - `Real` trait to constrain for `f64` and `Jet<N>`
+- Linear Algebra: effective `Matrix` structure, LU / QR / SVD / Cholesky decompositions (`O3` feature for the last three), determinant, inverse, block partitioning, reduced row echelon form, eigenvalue & eigenvector
+- Functional Programming: easier functional programming with `Vec<f64>`; matrix maps (`fmap`, `col_map`, `row_map`)
+- Automatic Differentiation: const-generic `Jet<N>` for arbitrary-order forward AD (`Dual`, `HyperDual` aliases), `#[ad_function]` proc macro, exact Jacobian via `jacobian()`, `Real` trait over `f64` and `Jet<N>`
 - Numerical Analysis
-  - Lagrange interpolation
-  - Splines
-    - Cubic Spline
-    - Cubic Hermite Spline
-      - Estimate slope via Akima
[…]
