### 2026-07-06 6d99f2f
DOCS: Document O3-accelerate Apple-only constraint (JOSS #98)
Enabling O3-accelerate on a non-Apple target fails while compiling
the accelerate-src dependency with an obscure rustc error
(library kind `framework` is only supported on Apple targets).
A guard inside the crate cannot improve this: the dependency is
compiled before peroxide itself, so neither a compile_error! in
lib.rs nor a build.rs check ever gets a chance to run (verified
empirically). Document the exact error string and the cargo-hack
exclude guidance in the Pre-requisite section instead, so users
hitting the error can find th
[…]
