### 2026-07-06 Axect/Peroxide#99 @Axect
@jonaspleyer Thanks for the follow-up, and apologies for the confusion my earlier summary caused. The per-file tree never actually landed in the README. Commit 86cc03d removed the old "Module Structure" section from the README and moved that content into `CONTRIBUTING.md` as a "Source layout" section, but my reply above described it the other way around. The current README contains: Quickstart, Why Peroxide?, Pre-requisite, Install, Examples, Release notes, Contributing, License, and Cite.

That said, your maintenance-burden point applies just as much to the tree in `CONTRIBUTING.md`, so I trimmed it down to a directory-level overview (one line per module with a short description). I also promoted the "Hello, Peroxide" example to a top-level Quickstart section right below the badges, so the first thing a new reader sees is runnable code. And the list-condensing format you suggested originally ("Probability Distributions (Bernoulli, Uniform, ...)") is now applied to the "Batteries included" section, which went from 89 lines to 15 without dropping any item. All three changes are in <https://github.com/Axect/Peroxide/commit/72a9e56>, already merged to master.

If the README looks good to you now, I believe this issue can be closed. Happy to adjust further if anything still feels redundant.
