### 2026-07-06 Axect/Peroxide#98 @Axect
@jonaspleyer Sorry, my earlier comment was ambiguous. To clarify: the crate does build and pass on my Arch machine. Running your script here, the only failing configurations were the three known ones: bare `O3` (needs a link backend), and `nc` / `netcdf` (HDF5 2.x vs the pinned `hdf5-sys 0.8`). Every combination that failed in your second run, including `blas,csv`, passed here. That is why I suspect an environment effect such as disk exhaustion rather than a code regression; one full pass of the matrix grew `target/` past 300 GB on my machine.

About the `cargo hack` error: ``library kind `framework` is only supported on Apple targets`` comes from `--each-feature` enabling `O3-accelerate`. Its `accelerate-src` backend emits an Apple-only framework link flag, so on Linux it must be excluded. The CI job runs exactly these commands:

```sh
cargo hack build --each-feature --optional-deps --keep-going \
  --exclude-features O3-accelerate,O3-mkl,O3-netlib,nc,netcdf,plot,pyo3,blas-src,lapack-src

cargo hack build --feature-powerset --depth 2 --optional-deps --keep-going \
  --exclude-features O3,O3-openblas,O3-accelerate,O3-mkl,O3-netlib,nc,netcdf,plot,pyo3,blas-src,lapack-src,arrow,parquet
```

These run on every push since fcfd012, so feature-gating regressions are now caught automatically.

On the additivity concern: the pure-Rust flags and the convenience `O3-*` flags are additive. The bare `O3` / `blas` / `lapack` flags are FFI-binding escape hatches that intentionally require the downstream binary to pick a link backend, and the README documents them as advanced flags. For the confusing `O3-accelerate` error I also tried adding a compile-time guard inside the crate, but it cannot help: the error occurs while compiling the `accelerate-src` dependency, before peroxide itself is ever compiled, so neither a `compile_error!` in `lib.rs` nor a `build.rs` check gets a chance to run (verified empirically). Instead, the exact error string, the platform constraint, and the cargo-hack exclude guidance are now documented in the README's Pre-requisite section (<https://github.com/Axect/Peroxide/commit/6d99f2f>, merged to master).

Looking forward to the logs from your re-run.
