### 2026-07-06 Axect/Peroxide#94 @Axect
Closing this out: the invitation period ended on April 26 and the author list has been finalized. Six contributors joined as co-authors (Giorgio Comitini, Jonas Grage, Benjamin Joens, Marc Schreiber, Soumya Sen, and Johanna Sörngård), with affiliations and ORCIDs reflected in paper.md and CITATION.cff on the features/joss-paper branch. Contributors who declined or did not respond are credited through the acknowledgements and the repository contributor history, as described above. Thanks to everyone who replied, and to all contributors for making Peroxide better. Status was reported to the JOSS review thread in openjournals/joss-reviews#10366.
