### 2026-07-06 febd4e2
FIX: Correct Gauss sum for even-order Gauss-Kronrod rules (#93)
The evaluation-reusing path (compute_gauss_kronrod_sums_stored,
introduced in faeaf2b) reconstructed the Gauss sum assuming an
odd-order Gauss rule: a center node at 0.0 plus pairs at Kronrod
offsets +/- 2j. Even-order rules (G10/G20/G30) have no node at 0.0;
their nodes sit at odd Kronrod offsets +/- (2j - 1). The old code
therefore added a spurious center term (a Kronrod-only node times an
ordinary pair weight), dropped one node pair, and read every pair
from Kronrod-only nodes, so the "Gauss" sum contained no Gauss node
at all.
[…]
