### 2026-07-06 2874984
SECURITY: Encapsulate Matrix / ComplexMatrix fields (#101)
Safe code could construct a Matrix whose data.len() disagrees with
row * col by writing the public fields directly, and the internal
unsafe code (raw pointers via ptr(), matrixmultiply / BLAS dgemm
calls, Index arithmetic) trusts those fields without checking the
buffer length, so heap out-of-bounds reads and writes were reachable
without any unsafe in user code (reported in #101 for peroxide
0.41.0).

Root fix, targeted at v0.42.0 (breaking change):

* Matrix and ComplexMatrix fields are now pub(crate); the invariant
  data.len() == r
[…]
