+++ tests/change_shape_test.rs
@@ -4,7 +4,7 @@ use peroxide::fuga::*;
 #[test]
 fn change_shape_test() {
     let a = matrix(vec![1, 2, 3, 4], 2, 2, Row);
-    assert_eq!(a.shape, Row);
+    assert_eq!(a.layout(), Row);
     let b = a.change_shape();
-    assert_eq!(b.shape, Col);
+    assert_eq!(b.layout(), Col);
 }
