### 2026-07-06 523185d
IMPL: Add trace, Hermitian conjugate, real/imag extractors (#87)
Add the small ergonomic operations requested in #87:

* Matrix::trace() and ComplexMatrix::trace(): sum of the diagonal,
  panics on non-square input (same contract as diag()).
* ComplexMatrix::h(): Hermitian conjugate (conjugate transpose).
* ComplexMatrix::real() / imag(): extract the real or imaginary part
  as a real Matrix, preserving the storage order.

The remaining items from #87 (IntoIterator support and a true matrix
exponential) are larger and stay tracked on the issue.
