### 2026-07-06 aa63524
FEAT: Make rand / rand_distr optional behind default-on rand feature (#88)
Sandboxed targets (Typst plugins, wasm32) cannot use stateful RNG, so
the random sampling stack is now optional:

* rand / rand_distr become optional dependencies behind a new
  default-on `rand` feature; existing users see no change, and
  `default-features = false` drops the RNG stack entirely.
* Gated behind the feature: statistics::dist, statistics::rand,
  util::wrapper, the rand() / rand_with_rng() / rand_with_dist()
  constructors, the runif! / rnorm! / dnorm! / pnorm! / rt! / dt! /
  pt! / rand! macros, the dist
[…]
