+++ src/fuga/mod.rs
@@ -183,10 +183,17 @@ pub use crate::complex::{integral::*, matrix::*, vector::*, C64};
 #[allow(ambiguous_glob_reexports)]
 pub use crate::structure::{ad::*, dataframe::*, matrix::*, polynomial::*, vector::*};
 
-pub use crate::util::{api::*, low_level::*, non_macro::*, print::*, useful::*, wrapper::*};
+pub use crate::util::{api::*, low_level::*, non_macro::*, print::*, useful::*};
+
+#[cfg(feature = "rand")]
+pub use crate::util::wrapper::*;
 
 #[allow(unused_imports)]
-pub use crate::statistics::{dist::*, ops::*, rand::*, stat::*};
+pub use crate::statistics::{ops::*, stat::*};
+
+#[cfg(feature = "rand")]
+#[allow(unused_imports)]
+pub use crate::statistics::{dist::*, rand::*};
 
 #[allow(unused_imports)]
 pub use crate::special::function::*;
@@ -205,6 +212,7 @@ pub use crate::util::plot::*;
 
 pub use anyhow;
 pub use paste;
+#[cfg(feature = "rand")]
 pub use rand::prelude::*;
 
 // =============================================================================
