### 2026-07-07 Axect/Peroxide#94 @russellb23
Unfortunately, I missed to read this message due to some technical issues accessing my github account. However,
Name: Russell R P Senthamarai
Affiliation: Independent researcher
ORCID: https://orcid.org/0000-0002-8061-5480
