### 2026-07-07 e494cf0
PAPER: Add Russell R P Senthamarai as co-author
Russell missed the invitation window on issue #94 due to GitHub access
issues and confirmed authorship afterward. He is a code contributor
(dot product for Dual numbers, with tests). Inserted in alphabetical
order by family name (between Sen and Sörngård) in paper.md and
CITATION.cff.
