+++ CITATION.cff
@@ -33,6 +33,10 @@ authors:
     family-names: Sen
     affiliation: Independent Researcher
     orcid: 'https://orcid.org/0009-0008-0337-9964'
+  - given-names: Russell R P
+    family-names: Senthamarai
+    affiliation: Independent Researcher
+    orcid: 'https://orcid.org/0000-0002-8061-5480'
   - given-names: Johanna
     family-names: Sörngård
     affiliation: Independent Researcher
