### 2026-07-07 Axect/Peroxide#94 @Axect
@russellb23 No problem at all, thanks for getting back to me despite the account trouble. Your contributions to Peroxide clearly warrant authorship, so I've added you to the paper. 👍

You're now in `paper.md` and `CITATION.cff` on the `features/joss-paper` branch. The paper is still under review at [openjournals/joss-reviews#10366](https://github.com/openjournals/joss-reviews/issues/10366), so this makes it in fine. Thanks again for your work on the library!
