### 2026-07-07 fbf5896
`cargo fmt`, unresolved imports, and `ln_gamma`
