### 2026-07-07 fa2470e
Improve numerical stability of `ln_gamma_approx` using Euler's reflection formula
