### 2026-07-07 59f985d
Handle poles
