### 2026-07-07 ef7068b
Reimplement factorial fast-path
