+++ src/special/lanczos.rs
@@ -1,6 +1,6 @@
 //! Lanczos approximation Coefficient generator
 
-use crate::statistics::ops::{double_factorial, factorial, C};
+use crate::statistics::ops::{double_factorial, C};
 use crate::structure::matrix::Matrix;
 use crate::traits::matrix::MatrixTrait;
 use crate::traits::pointer::{Oxide, RedoxCommon};
@@ -50,7 +50,7 @@ pub fn gamma_approx(z: f64) -> f64 {
         }
     }
 
-    if z > 1.0 && z.fract() == 0.0 {
+    if z > 0.0 && z.fract() == 0.0 {
         let mut result = 1.0;
         let n = (z - 1.0) as u64;
         for i in 2..=n {
