+++ src/statistics/dist.rs
@@ -241,6 +241,7 @@
 //!     * Mean: $e^{\mu + \frac{\sigma^2}{2}}$
 //!     * Var: $(e^{\sigma^2} - 1)e^{2\mu + \sigma^2}$
 //! * To generate log-normal random samples, Peroxide uses the `rand_distr::LogNormal` distribution from the `rand_distr` crate.
+//!
 //! ### `MVRNG` trait
 //!
 //! * `MVRNG` trait is composed of four fields
@@ -270,8 +271,8 @@
 //! ### Dirichlet Distribution
 //!
 //! * Definition
-//!     $$ \text{Dir}(\mathbf{x} | \boldsymbol{\alpha}) = \frac{1}{\text{B}(\boldsymbol{\alpha})} \prod_{i=1}^K x_i^{\alpha_i - 1} $$
-//!     where $\text{B}(\boldsymbol{\alpha}) = \frac{\prod_{i=1}^K \Gamma(\alpha_i)}{\Gamma(\sum_{i=1}^K \alpha_i)}$
+//!   $$ \text{Dir}(\mathbf{x} | \boldsymbol{\alpha}) = \frac{1}{\text{B}(\boldsymbol{\alpha})} \prod_{i=1}^K x_i^{\alpha_i - 1} $$
+//!   where $\text{B}(\boldsymbol{\alpha}) = \frac{\prod_{i=1}^K \Gamma(\alpha_i)}{\Gamma(\sum_{i=1}^K \alpha_i)}$
 //! * Representative value
 //!     * Mean: $\frac{\alpha_i}{\alpha_0}$
 //!     * Var : $\frac{\alpha_i(\alpha_0 - \alpha_i)}{\alpha_0^2(\alpha_0 + 1)}$
@@ -292,6 +293,68 @@
 //!         a.cov().print();                            // Covariance matrix
 //!     }
 //!     ```
+//!
+//! ### `MatRNG` trait
+//!
+//! * `MatRNG` trait is composed of four fields for Matrix-Variate distributions
+//!     * `sample`: Extract samples
+//!     * `sample_with_rng`: Extract samples with specific rng
+//!     * `pdf` : Calculate pdf value at specific point
+//!     * `ln_pdf` : Calculate log-pdf value at specific point
+//!     ```no_run
+//!     use rand::Rng;
+//!     use peroxide::fuga::*;
+//!
+//!     pub trait MatRNG {
+//!         /// Extract samples of matrix-variate distributions
+//!         fn sample(&self, n: usize) -> Vec<Matrix>;
+//!
+//!         /// Extract samples of distributions with specific rng
+//!         fn sample_with_rng<R: Clone + Rng>(&self, rng: &mut R, n: usize) -> Vec<Matrix>;
+//!
+//!         /// Probability Density Function
+//!         fn pdf(&self, x: &Matrix) -> f64;
+//!
+//!         /// Log Probability Density Function
+//!         fn ln_pdf(&self, x: &Matrix) -> f64;
+//!     }
+//!     ```
+//!
+//! ### Wishart Distribution
+//!
+//! * Definition
+//!   $$ \mathcal{W}_p(\nu, \mathbf{V}) = \frac{|\mathbf{X}|^{(\nu - p - 1)/2} \exp\left(-\frac{1}{2}\text{tr}(\mathbf{V}^{-1}\mathbf{X})\right)}{2^{\nu p / 2} |\mathbf{V}|^{\nu/2} \Gamma_p\left(\frac{\nu}{2}\right)} $$
[…]
