### 2026-07-08 Axect/Peroxide#98 @jonaspleyer
Okay I understand. I think that this all sounds reasonable overall. I have manually verified some of the tests and you are probably right in that it was an environment problem (probably disk space).

However, as I have written earlier I tried to follow the guidelines for installing the additional dependencies on my Arch machine which are explained here https://github.com/Axect/Peroxide_BLAS and although I have now managed to compile `peroxide` with the `blas` feature, the way of how to do this seems overly complicated.
What is the reason that the `openblas-src` crate is not an optional dependency of `peroxide` which is activated by the `blas` feature?
Furthermore I have found that `openblas-src` requires either `rustls` or `native-tls` as an additional feature which is not properly documented. Maybe there is a good reason for the way that it is right now. But to me it seems overly complicated and slightly outdated.
The documentation of this feature should not reside in a separate github repo imho.
