### 2026-07-08 Axect/Peroxide#99 @jonaspleyer
Do you think that trimming it down in the `CONTRIBUTING.md` solves the problem?
