### 2026-07-09 2f69f51
add bidiagonalize method to matrix
