### 2026-07-09 934730f
add givens rotation method to matrix
