### 2026-07-10 6fdc958
IMPL: Add O3-openblas-system feature and correct BLAS setup docs (JOSS jonaspleyer #98)
- New O3-openblas-system feature links the system-installed OpenBLAS via
  pkg-config instead of compiling OpenBLAS from source; openblas-src is
  added as an optional dependency solely to forward its system feature.
- README Pre-requisite section now states that O3-openblas builds OpenBLAS
  from source (C/Fortran toolchain, make, network access) and documents the
  rustls / native-tls requirement of openblas-src 0.10.16+ when used with
  default-features = false.
- Drop the stale Peroxide_BLAS link from t
[…]
