### 2026-07-10 Axect/Peroxide#98 @Axect
@jonaspleyer Thanks for following up. All three points are addressed in #107.

**On making `openblas-src` an optional dependency of the `blas` feature.**

- It effectively already is, one level up: since v0.41.1, `O3-openblas` pulls `openblas-src` in through `blas-src/openblas` + `lapack-src/openblas`. `cargo add peroxide --features O3-openblas` needs nothing else in the downstream `Cargo.toml`.
- The bare `blas` / `O3` flags stay backend-free deliberately: cargo features are additive, so if `blas` hard-wired OpenBLAS, a downstream project that wants BLIS or R's BLAS could never opt out. This is the same bindings/backend split used across the Rust BLAS ecosystem (`ndarray-linalg` has the analogous `openblas-*` / `intel-mkl-*` family).
- The root cause of the complexity you experienced is the outdated README link: it pointed to the Peroxide_BLAS repo, which describes the manual setup for peroxide 0.33 (`default-features = false`, `extern crate`) and predates the `O3-openblas` flag. Following it leads to the manual path even though the convenience flag exists. #107 removes the link, and I will archive that repo with a deprecation notice, so the feature documentation lives in exactly one place.

**On `rustls` / `native-tls`.** Confirmed, and you are right that it was undocumented.

- `openblas-src` 0.10.16 (2026-05) started downloading the OpenBLAS source tarball over HTTPS via `ureq`; the TLS backend is chosen by a `rustls` / `native-tls` feature.
- `rustls` is included in its default feature set, so the standard `O3-openblas` path is unaffected. The failure only occurs with `default-features = false`, which is the configuration the old guide recommended.
- The Pre-requisite section now documents this explicitly.

**While fixing this I found an error in the README about what `O3-openblas` requires.**

- It claimed system OpenBLAS packages are required. They are not: without the `system` feature, `openblas-src` always compiles OpenBLAS from source. The actual requirements are `gcc` / `gfortran` / `make` and network access, not `libopenblas`. The backend table now lists the actual build-time requirements per flag.
- #107 adds `O3-openblas-system`, which links the system-installed OpenBLAS via pkg-config instead of building from source. On Arch this is likely the better fit: `pacman -S openblas`, then `cargo add peroxide --features O3-openblas-system`.
- Verified here against openblas 0.3.33 (483 tests pass); CI now builds and tests both variants.

If this addresses the remaining points, I believe this issue can be closed once #107 is merged. Happy to adjust the docs further if anything still seems overly complicated.
