### 2026-07-10 737611c
DOCS: Replace source layout table with pointer to docs.rs module index (JOSS jonaspleyer #99)
The hand-maintained module table duplicated the module index that rustdoc
generates on docs.rs and could only drift out of sync. What remains is the
repo-only fact rustdoc cannot show: src/grave/ is retired code excluded
from compilation.
