### 2026-07-10 ee6030d
DOCS: Move module descriptions from CONTRIBUTING table into module docs
The removed source layout table carried descriptions that existed nowhere
in rustdoc: complex and traits had no module docs at all, and several
first lines (Useful macros, Statistical Modules, Main structures) were too
vague to serve as an index. Each module now carries its one-line purpose
in its own mod.rs, so the docs.rs module index is the single source for
the source map and stays in sync with the code by construction.
