### 2026-07-10 Axect/Peroxide#99 @Axect
@jonaspleyer No, you are right, it does not. A trimmed table is still a hand-maintained mirror of the source structure, and any curated copy can only drift.

But when I went to delete it, it turned out the table was not fully redundant: `complex` and `traits` had no module docs at all, and several module doc first lines ("Useful macros", "Statistical Modules") were too vague to serve as an index. Deleting the table alone would have removed information that existed nowhere else.

So #108 relocates instead of deletes, following the same principle you suggested for the README (keep the information in the cargo doc output, not in curated markdown):

- Each module's one-line purpose now lives in its own `mod.rs` (`//!` docs), so the module index on docs.rs is the source map, and the descriptions are reviewed together with the code they describe.
- CONTRIBUTING.md keeps only a two-sentence pointer plus the one repo-only fact rustdoc cannot show: `src/grave/` holds retired implementations excluded from compilation.

There is no separate structure document left to fall out of sync. I believe this resolves the last open point of this issue; feel free to close once #108 is merged.
