### 2026-07-10 Axect/Peroxide#98 @jonaspleyer
What is your plan with the other repo then? Also: It seems that you are generating these responses with AI. I would prefer to write directly with you. Letting AI code is fine to me. But I prefer human communication.
