### 2026-07-10 Axect/Peroxide#99 @jonaspleyer
That is exactly what `rustdoc` was developed for.
