### 2026-07-11 ce2a1c3
Add magnitude guard
