### 2026-07-11 Axect/Peroxide#98 @Axect
Thank you for the comments. And I'm sorry for making you feel uncomfortable. There are so many things I should update, and it's quite tough to summarize, so I borrow AI's capability to summarize updates. (But all modifications are tested by myself)

I archived Peroxide_BLAS repo. It is too outdated. And I already left deprecated warning on the README.md of the repo. You can check the repo [Peroxide_BLAS](https://github.com/Axect/Peroxide_BLAS).
