### 2026-07-11 24fb64f
Ease feature restriction on `ln_pdf`, deref in Print impl, add more tests
