### 2026-07-11 d76617f
Add covariance convergence test
