### 2026-07-11 c004d04
Remove `cfg` on `mv_ln_gamma`
