### 2026-07-12 05a2b53
Imrpoves docs and notation
