### 2026-07-12 1ea5037
Add wilkinson_shift method to matrix
