### 2026-07-21 Axect/Peroxide#98 @jonaspleyer
Marking the repo as outdated is good. Where can users now find the information in the main repo? Are you planning to add more information as users from the community share their setup on which packages to install for their OS/linux distribution etc.?
