### 2026-07-28 Axect/Peroxide#98 @Axect
I moved that information to the [Pre-requisite section](https://github.com/Axect/Peroxide#pre-requisite) in README.md. I mentioned the difference between `O3-openblas` and `O3-openblas-system` there, and also the per-distro install commands, and the Apple-only error from `O3-accelerate`. Nothing is lost from `Peroxide_BLAS`. And also, I included the information about which packages should be installed for different OS and distros in same section. (if they choose `O3-openblas-system`.)
