### 2026-08-02 645aa79
PAPER: Refresh countable claims and the Zenodo reference entry
Full read requested by the handling editor before recommend-accept
(openjournals/joss-reviews#10366). Four numbers had drifted:

- The `peroxide` bibliography entry still described the v0.41.0 deposit
  with a single author and the old Zenodo title. The archived record
  now carries all eight authors under the manuscript title, so the
  entry is updated to match what the DOI resolves to.
- Standalone examples went from 41 to 42 in v0.42.0.
- "A test suite of 26 modules" matched no measurable quantity. The
  repository has 22 integr
[…]
