+++ paper/paper.bib
@@ -1,8 +1,8 @@
 % ===== Peroxide itself =====
 @software{peroxide,
-  author       = {Kim, Tae-Geun},
-  title        = {{Peroxide: Rust numeric library containing linear algebra, numerical analysis, statistics, and machine learning tools}},
-  version      = {0.41.0},
+  author       = {Kim, Tae-Geun and Comitini, Giorgio and Grage, Jonas and Joens, Benjamin and Schreiber, Marc and Sen, Soumya and Senthamarai, Russell R P and S\"{o}rng\r{a}rd, Johanna},
+  title        = {{Peroxide: A Batteries-Included Numerical Computing Library for Rust}},
+  version      = {0.43.1},
   year         = {2026},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.10815823},
