### 2026-08-09 82a48b3
Update paper.bib
