+++ paper/paper.bib
@@ -1,7 +1,7 @@
 % ===== Peroxide itself =====
 @software{peroxide,
   author       = {Kim, Tae-Geun and Comitini, Giorgio and Grage, Jonas and Joens, Benjamin and Schreiber, Marc and Sen, Soumya and Senthamarai, Russell R P and S\"{o}rng\r{a}rd, Johanna},
-  title        = {{Peroxide: A Batteries-Included Numerical Computing Library for Rust}},
+  title        = {Peroxide: A Batteries-Included Numerical Computing Library for {R}ust},
   version      = {0.43.1},
   year         = {2026},
   publisher    = {Zenodo},
@@ -257,7 +257,7 @@
 
 @software{russell,
   author  = {Pedroso, Dorival M.},
-  title   = {Russell: {Rust} Scientific Library},
+  title   = {{r}ussell: {Rust} Scientific Library},
   url     = {https://github.com/cpmech/russell},
   version = {1.10.0},
   year    = {2025}
